% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linnet.R
\name{as.linnet}
\alias{as.linnet}
\alias{as.linnet.sfnetwork}
\title{Convert an sfnetwork object into linnet format}
\usage{
as.linnet.sfnetwork(X, ...)
}
\arguments{
\item{X}{An \code{sfnetwork} object with implicit or explicit edges and
projected CRS.}

\item{...}{Arguments passed to \code{\link[spatstat]{linnet}} (only
\code{sparse} and \code{warn}).}
}
\value{
An object of class \code{linnet}.
}
\description{
A method to convert \code{sfnetwork} object into
\code{\link[spatstat]{linnet}} format and enhance the interoperability
between \code{sfnetworks} and \code{spatstat}. Use this method without the
.sfnetwork suffix and after loading the \code{spatstat} package.
}
\details{
Check also \code{\link[sfnetworks]{as_sfnetwork}} for other methods to
convert \code{\link[spatstat]{linnet}} objects into \code{sfnetwork} format.
}
\examples{
if (require("spatstat", quietly = TRUE)) {
  roxel_sfn = as_sfnetwork(roxel) \%>\%
    sf::st_transform(3035)
  (roxel_linnet = as.linnet(roxel_sfn, sparse = TRUE))
  plot(roxel_linnet)
}
}
