% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_points}
\alias{as_points}
\title{Represent any geometry as a point}
\usage{
as_points(x)
}
\arguments{
\item{x}{An object of class \code{\link[sf]{sf}}.}
}
\description{
Represent any geometry as a point
}
\details{
Centroids will be calculated for geometries that are not points.
}
