% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{as_s2_geography}
\alias{as_s2_geography}
\alias{as_s2_geography.sfnetwork}
\title{Extract the geometries of a sfnetwork as a S2 geography vector}
\usage{
as_s2_geography.sfnetwork(x, focused = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{...}{Arguments passed on the corresponding \code{s2} function.}
}
\value{
An object of class \code{\link[s2]{s2_geography}}.
}
\description{
A method to convert an object of class \code{\link{sfnetwork}} into
\code{\link[s2]{s2_geography}} format. Use this method without the
.sfnetwork suffix and after loading the \pkg{s2} package.
}
