% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot}
\alias{autoplot}
\alias{autoplot.sfnetwork}
\title{autoplot method for sfnetworks}
\usage{
autoplot.sfnetwork(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{sfnetwork}}.}

\item{...}{Ignored.}
}
\description{
Plot the geometries of an object of class \code{\link{sfnetwork}} 
automatically as a \code{\link[ggplot2]{ggplot}} object. Use this method 
without the .sfnetwork suffix and after loading the ggplot2 package.
}
\details{
See \code{\link[ggplot2]{autoplot}}.
}
\examples{
library(ggplot2)
library(sf)
net = as_sfnetwork(roxel)

# Quick overview of the network in `ggplot2` style
autoplot(net)

# Other `ggplot2` elements can be added
points = net \%>\%
  st_bbox() \%>\%
  st_as_sfc() \%>\%
  st_sample(10, type = 'random') \%>\%
  st_set_crs(4326) \%>\%
  st_cast('POINT')

autoplot(net) +
   # The theme can be customized
   theme_minimal() +
   # Labels can be added
   labs(title = 'Nice ggplot') +
   # And extra `geom_sf` layers can be included
   geom_sf(data = points, color = 'red', size = 2)

}
