% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.sfnetwork}
\alias{autoplot.sfnetwork}
\title{Autoplot sfnetwork object}
\usage{
autoplot.sfnetwork(x)
}
\arguments{
\item{x}{Object of class \code{\link{sfnetwork}}.}
}
\description{
Autoplot an object of class \code{\link{sfnetwork}} as a ggplot object.
}
\details{
This is a basic ggploting functionality, for a quick assessment of
the network but on a \code{ggplot2} format.
}
\examples{

net = as_sfnetwork(roxel)

# Quick overview of the network in `ggplot2` style
autoplot(net)

# Other `ggplot2` elements can be added
points = net \%>\%
  st_bbox() \%>\%
  st_as_sfc() \%>\%
  st_sample(10, type = 'random') \%>\%
  st_set_crs(4326) \%>\%
  st_cast('POINT')

autoplot(net) +
   # The theme can be customized
   ggplot2::theme_minimal() +
   # Labels cna be added
   ggplot2::labs(title = 'Nice ggplot') +
   # And extra `geom_sf` layers can be included
   ggplot2::geom_sf(data = points, color = 'red', size = 2)

}
