% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind_spatial}
\alias{bind_spatial}
\alias{bind_spatial_nodes}
\alias{bind_spatial_edges}
\title{Add nodes or edges to a spatial network.}
\usage{
bind_spatial_nodes(.data, ...)

bind_spatial_edges(.data, ..., node_key = "name", force = FALSE)
}
\arguments{
\item{.data}{An object of class \code{\link{sfnetwork}}.}

\item{...}{One or more objects of class \code{\link[sf]{sf}} containing the
nodes or edges to be added.}

\item{node_key}{The name of the column in the nodes table that character
represented \code{to} and \code{from} columns should be matched against. If
\code{NA}, the first column is always chosen. This setting has no effect if
\code{to} and \code{from} are given as integers. Defaults to \code{'name'}.}

\item{force}{Should network validity checks be skipped? Defaults to
\code{FALSE}, meaning that network validity checks are executed after binding
edges, making sure that boundary points of edges match their corresponding
node coordinates.}
}
\value{
An object of class \code{\link{sfnetwork}} with added nodes or
edges.
}
\description{
These functions are the spatially aware versions of tidygraph's
\code{\link[tidygraph]{bind_nodes}} and \code{\link[tidygraph]{bind_edges}}
that allow you to add rows to the nodes or edges tables in a
\code{\link{sfnetwork}} object. As with \code{\link[dplyr]{bind_rows}}
columns are matched by name and filled with \code{NA} if the column does not
exist in some instances.
}
\examples{
library(sf, quietly = TRUE)
library(dplyr, quietly = TRUE)

net = roxel |>
  slice(c(1:2)) |>
  st_transform(3035) |>
  as_sfnetwork()

pts = roxel |>
  slice(c(3:4)) |>
  st_transform(3035) |>
  st_centroid()

bind_spatial_nodes(net, pts)

}
