% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capture_cat}
\alias{capture_cat}
\title{Subset and modify the result of \code{capture.output} to pass onto print method.}
\usage{
capture_cat(p, start, end, prefix = "", sep = "\\n", style = "plain")
}
\arguments{
\item{p}{The result of \code{capture.output}.}

\item{start}{From which line to start subsetting.}

\item{end}{From which line to finish subsetting.}

\item{prefix}{A string that should be passed before printing each line.}

\item{sep}{How to separate each line for printing.}

\item{style}{Whic style to print to console. Options are 'plain' (default),
'subtle' which makes it silver color, and 'subtle_italic' which makes it
silver color and italic.}
}
\value{
A modified subset of the original captured output.
}
\description{
Subset and modify the result of \code{capture.output} to pass onto print method.
}
