% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cat_subtle}
\alias{cat_subtle}
\title{Print a string with a subtle style}
\usage{
cat_subtle(...)
}
\arguments{
\item{...}{A string to print}
}
\value{
A printed string to console with subtle style
}
\description{
Print a string with a subtle style
}
