% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_nodes_from_edges}
\alias{create_nodes_from_edges}
\title{Create nodes from edges}
\usage{
create_nodes_from_edges(edges)
}
\arguments{
\item{edges}{An object of class \code{\link[sf]{sf}} with \code{LINESTRING}
geometries.}
}
\value{
A list with the edges as an object of class \code{\link[sf]{sf}}
with \code{LINESTRING} geometries and the created nodes as an object of
class \code{\link[sf]{sf}} with \code{POINT} geometries.
}
\description{
Create nodes from edges
}
\details{
It is assumed that the given LINESTRING geometries form the edges.
Nodes need to be created at the endpoints of the edges. Identical endpoints
need to be the same node.
}
