% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data}
\alias{data}
\alias{node_data}
\alias{edge_data}
\title{Extract the node or edge data from a spatial network}
\usage{
node_data(x, focused = TRUE)

edge_data(x, focused = TRUE, require_sf = FALSE)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{focused}{Should only features that are in focus be extracted? Defaults
to \code{TRUE}. See \code{\link[tidygraph]{focus}} for more information on
focused networks.}

\item{require_sf}{Is an \code{\link[sf]{sf}} object required? This will make
extraction of edge data fail if the edges are spatially implicit. Defaults
to \code{FALSE}.}
}
\value{
For the nodes, always an object of class \code{\link[sf]{sf}}. For
the edges, an object of class \code{\link[sf]{sf}} if the edges are
spatially explicit, and an object of class \code{\link[tibble]{tibble}}
if the edges are spatially implicity and \code{require_sf = FALSE}.
}
\description{
Extract the node or edge data from a spatial network
}
