% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draw_lines}
\alias{draw_lines}
\title{Draw lines between two sf objects, row-wise}
\usage{
draw_lines(x, y)
}
\arguments{
\item{sources}{An object of class \code{\link[sf]{sf}} with \code{POINT}
geometries, representing the points where lines need to start from.}

\item{targets}{An object of class \code{\link[sf]{sf}} with \code{POINT}
geometries, representing the points where lines need to end at.}
}
\value{
An object of class \code{\link[sf]{sfc}} with \code{LINESTRING}
geometries.
}
\description{
Draw lines between two sf objects, row-wise
}
\details{
Lines are drawn row-wise. That is, between point 1 in x and point 1
in y, point 2 in x and point 2 in y, et cetera.
}
