% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_geometry}
\alias{drop_geometry}
\title{Drop the geometry of a network element}
\usage{
drop_geometry(x, what)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}}

\item{what}{Either 'nodes' or 'edges'.}
}
\value{
An object of class \code{\link{sfnetwork}} when what = 'edges', and
an object of class \code{\link[tidygraph]{tbl_graph}} when what = 'nodes'.
}
\description{
Drop the geometry of a network element
}
