% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{dual_weights}
\alias{dual_weights}
\title{Specify dual edge weights}
\usage{
dual_weights(reported, actual)
}
\arguments{
\item{reported}{The edge weights to be reported. Evaluated by
\code{\link{evaluate_weight_spec}}.}

\item{actual}{The actual edge weights to be used to determine shortest paths.
Evaluated by \code{\link{evaluate_weight_spec}}.}
}
\value{
An object of class \code{dual_weights}.
}
\description{
Dual edge weights are two sets of edge weights, one (the actual weight) to
determine the shortest path, and the other (the reported weight) to report
the cost of that path.
}
\details{
Dual edge weights enable dual-weighted routing. This is supported
by the \code{\link[dodgr]{dodgr}} routing backend.
}
