% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{evaluate_node_query}
\alias{evaluate_node_query}
\title{Query specific node indices from a spatial network}
\usage{
evaluate_node_query(data, query)
}
\arguments{
\item{data}{An object of class \code{\link{sfnetwork}}.}

\item{query}{The query that defines for which nodes to extract indices,
defused into a \code{\link[dplyr:topic-quosure]{quosure}}. See Details for
the different ways in which node queries can be formulated.}
}
\value{
A vector of queried node indices.
}
\description{
This function is not meant to be called directly, but used inside other
functions that accept a node query.
}
\details{
There are multiple ways in which node indices can be queried in
sfnetworks. The query can be formatted as follows:

\itemize{
  \item As spatial features: Spatial features can be given as object of
  class \code{\link[sf]{sf}} or \code{\link[sf]{sfc}}. The nearest node to
  each feature is found by calling \code{\link[sf]{st_nearest_feature}}.
  \item As node type query function: A
  \link[tidygraph:node_types]{node type query function} defines for each
  node if it is of a given type or not. Nodes that meet the criterium are
  queried.
  \item As node predicate query function: A
  \link[=spatial_node_predicates]{node predicate query function} defines
  for each node if a given spatial predicate applies to the spatial relation
  between that node and other spatial features. Nodes that meet the
  criterium are queried.
  \item As column name: The referenced column is expected to have logical
  values defining for each node if it should be queried or not. Note that
  tidy evaluation is used and hence the column name should be unquoted.
  \item As integers: Integers are interpreted as node indices. A node index
  corresponds to a row-number in the nodes table of the network.
  \item As characters: Characters are interpreted as node names. A node name
  corresponds to a value in a column named "name" in the the nodes table of
  the network. Note that this column is expected to store unique names
  without any duplicated values.
  \item As logicals: Logicals should define for each node if it should be
  queried or not.
}

Queries that can not be evaluated in any of the ways described above will be
forcefully converted to integers using \code{\link{as.integer}}.
}
