% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_edges}
\alias{get_edges}
\title{Point to the edges in an sfnetwork}
\usage{
get_edges(x)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}
}
\value{
An object of class \code{\link{sfnetwork}} with activated edges.
}
\description{
Point to the edges in an sfnetwork
}
