% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_endpoint_coords}
\alias{get_endpoint_coords}
\title{Get the X or Y coordinates of the endpoints of LINESTRING geometries}
\usage{
get_endpoint_coords(x, XY)
}
\arguments{
\item{x}{An object of class \code{\link[sf]{sf}} with \code{LINESTRING}
geometries.}

\item{XY}{Which coordinates do you want to get? Either "X" or "Y".}
}
\description{
Get the X or Y coordinates of the endpoints of LINESTRING geometries
}
