% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_endpoints}
\alias{get_endpoints}
\title{Get the endpoints of LINESTRING geometries}
\usage{
get_endpoints(x)
}
\arguments{
\item{x}{An object of class \code{\link[sf]{sf}} with \code{LINESTRING}
geometries.}
}
\value{
An object of class \code{data.frame} with an "X" and "Y" column,
containing respectively the X and Y coordinates of the endpoints.
}
\description{
Get the endpoints of LINESTRING geometries
}
