% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_geometry_colname}
\alias{get_geometry_colname}
\title{Retrieve the name of the geometry list column in an sf object}
\usage{
get_geometry_colname(x)
}
\arguments{
\item{x}{An object of class \code{\link[sf]{sf}}.}
}
\value{
The name of the geometry list column as a string.
}
\description{
Retrieve the name of the geometry list column in an sf object
}
