% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nodes}
\alias{get_nodes}
\title{Point to the nodes in an sfnetwork}
\usage{
get_nodes(x)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}
}
\value{
An object of class \code{\link{sfnetwork}} with activated nodes.
}
\description{
Point to the nodes in an sfnetwork
}
