% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group_spatial}
\alias{group_spatial}
\title{Group nodes based on spatial distance}
\usage{
group_spatial(
  dist,
  algorithm = "dbscan",
  network_distance = TRUE,
  min_nodes = 1,
  ...
)
}
\arguments{
\item{dist}{distance within which nodes are clustered to each other}

\item{algorithm}{the grouping algorithm used to perform spatial clustering.
Defaults to the `dbscan::dbscan()` algorithm.}

\item{network_distance}{should the distance be based on the network distance
(default, uses `st_network_distance()` internally) or euclidean distance
(uses `sf::st_distance()` internally)?}

\item{min_nodes}{minimum number of nodes assigned to each cluster. Defaults
to 1 so that every node is assigned a cluster even if it is the only member
of that cluster.}

\item{...}{other arguments passed onto the algorithm,
e.g. `dbscan::dbscan()`}
}
\description{
Group nodes based on spatial distance
}
