% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_spatially_explicit_edges}
\alias{has_spatially_explicit_edges}
\title{Check if an sfnetwork has spatially explicit edges}
\usage{
has_spatially_explicit_edges(x)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}
}
\value{
\code{TRUE} if the network has spatially explicit edges, \code{FALSE}
otherwise.
}
\description{
Check if an sfnetwork has spatially explicit edges
}
