% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfnetwork.R
\name{is.sfnetwork}
\alias{is.sfnetwork}
\title{Check if an object is an sfnetwork}
\usage{
is.sfnetwork(x)
}
\arguments{
\item{x}{Object to be checked.}
}
\description{
Check if an object is an sfnetwork
}
\examples{
library(tidygraph, quietly = TRUE)

net = as_sfnetwork(roxel)
is.sfnetwork(net)
is.sfnetwork(as_tbl_graph(net))

}
