% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_directed}
\alias{is_directed}
\title{Check if a graph is directed.}
\usage{
is_directed(x)
}
\arguments{
\item{x}{An object of \code{\link{sfnetwork}} or \code{tbl_graph}.}
}
\value{
\code{TRUE} when the given graph is directed, \code{FALSE} otherwise.
}
\description{
Check if a graph is directed.
}
