% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_spatially_explicit}
\alias{is_spatially_explicit}
\title{Check if a table has spatial information stored in a geometry list column}
\usage{
is_spatially_explicit(x)
}
\arguments{
\item{x}{Object to check for spatial explicitness}
}
\value{
\code{TRUE} if the table has a geometry list column, \code{FALSE}
otherwise.
}
\description{
Check if a table has spatial information stored in a geometry list column
}
