% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.R
\name{make_edges_directed}
\alias{make_edges_directed}
\title{Convert undirected edges into directed edges based on their geometries}
\usage{
make_edges_directed(x)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}
}
\value{
A directed network as object of class \code{\link{sfnetwork}}.
}
\description{
This function converts an undirected network to a directed network following
the direction given by the linestring geometries of the edges.
}
\details{
In undirected spatial networks it is required that the boundary of
edge geometries contain their incident node geometries. However, it is not
required that their start point equals their specified *from* node and their
end point their specified *to* node. Instead, it may be vice versa. This is
because for undirected networks *from* and *to* indices are always swopped
if the *to* index is lower than the *from* index. Therefore, the direction
given by the *from* and *to* indices does not necessarily match the
direction given by the edge geometries.
}
\note{
If the network is already directed it is returned unmodified.
}
