% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.R
\name{make_edges_explicit}
\alias{make_edges_explicit}
\title{Construct edge geometries for spatially implicit networks}
\usage{
make_edges_explicit(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{...}{Arguments forwarded to \code{\link[sf]{st_as_sf}} to directly
convert the edges table into a sf object. If no arguments are given, the
edges are made explicit by simply drawing straight lines between the start
and end node of each edge.}
}
\value{
An object of class \code{\link{sfnetwork}} with spatially explicit
edges.
}
\description{
This function turns spatially implicit networks into spatially explicit
networks by adding a geometry column to the edge data.
}
\note{
If the network is already spatially explicit it is returned
unmodified.
}
