% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.R
\name{make_edges_implicit}
\alias{make_edges_implicit}
\title{Drop edge geometries of spatially explicit networks}
\usage{
make_edges_implicit(x)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}
}
\value{
An object of class \code{\link{sfnetwork}} with spatially implicit
edges.
}
\description{
This function turns spatially explicit networks into spatially implicit
networks by dropping the geometry column of the edge data.
}
\note{
If the network is already spatially implicit it is returned
unmodified.
}
