% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.R
\name{make_edges_mixed}
\alias{make_edges_mixed}
\title{Make some edges directed and some undirected}
\usage{
make_edges_mixed(x, directed)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{directed}{An integer vector of edge indices specifying those edges
that should be directed.}
}
\value{
A mixed network as object of class \code{\link{sfnetwork}}.
}
\description{
This function creates a mixed network, meaning that some edges are directed,
and some are undirected. In practice this is implemented as a directed
network in which those edges that are meant to be undirected are duplicated
and reversed.
}
