% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.R
\name{make_edges_valid}
\alias{make_edges_valid}
\title{Match edge geometries to their incident node locations}
\usage{
make_edges_valid(x, preserve_geometries = FALSE)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{preserve_geometries}{Should the edge geometries remain unmodified?
Defaults to \code{FALSE}. See Details.}
}
\value{
An object of class \code{\link{sfnetwork}} with corrected edge
geometries.
}
\description{
This function makes invalid edges valid by modifying either edge or node
geometries such that the boundary points of edge linestring geometries
always match the point geometries of the nodes that are specified as their
incident nodes by the *from* and *to* columns.
}
\details{
If geometries should be preserved, edges are made valid by adding
edge boundary points that do not equal their corresponding node geometry as
new nodes to the network, and updating the *from* and *to* indices to match
this newly added nodes. If \code{FALSE}, edges are made valid by modifying
their geometries, i.e. edge boundary points that do not equal their
corresponding node geometry are replaced by that node geometry.
}
\note{
This function works only if the edge geometries are meant to start at
their specified *from* node and end at their specified *to* node. In
undirected networks this is not necessarily the case, since edge geometries
are allowed to start at their specified *to* node and end at their specified
*from* node. Therefore, in undirected networks those edges first have to be
reversed before running this function. Use
\code{\link{make_edges_follow_indices}} for this.
}
