% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiple_matches}
\alias{multiple_matches}
\title{Check if the output of an st_join operation has multiple matches}
\usage{
multiple_matches(x)
}
\arguments{
\item{x}{The output of an st_join(a,b) operation where the original row
indices of a are stored in a column names \code{.sfnetwork_index}.}
}
\value{
\code{TRUE} when there where multiple matches, \code{FALSE}
otherwise.
}
\description{
Check if the output of an st_join operation has multiple matches
}
