% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb.R
\name{nb}
\alias{nb}
\alias{nb_to_sfnetwork}
\alias{sfnetwork_to_nb}
\title{Conversion between neighbor lists and sfnetworks}
\usage{
nb_to_sfnetwork(
  x,
  nodes,
  directed = TRUE,
  edges_as_lines = TRUE,
  compute_length = FALSE
)

sfnetwork_to_nb(x, direction = "out")
}
\arguments{
\item{x}{For the conversion to sfnetwork: a neighbor list, which is a list
adjacent to. For the conversion from sfnetwork: an object of class
\code{\link{sfnetwork}}.}

\item{nodes}{The nodes themselves as an object of class \code{\link[sf]{sf}}
or \code{\link[sf]{sfc}} with \code{POINT} geometries.}

\item{directed}{Should the constructed network be directed? Defaults to
\code{TRUE}.}

\item{edges_as_lines}{Should the created edges be spatially explicit, i.e.
have \code{LINESTRING} geometries stored in a geometry list column? Defaults
to \code{TRUE}.}

\item{compute_length}{Should the geographic length of the edges be stored in
a column named \code{length}? Defaults to \code{FALSE}.}

\item{direction}{The direction that defines if two nodes are neighbors.
Defaults to \code{'out'}, meaning that the direction given by the network is
followed and node j is only a neighbor of node i if there exists an edge
i->j. May be set to \code{'in'}, meaning that the opposite direction is
followed and node j is only a neighbor of node i if there exists an edge
j->i. May also be set to \code{'all'}, meaning that the network is
considered to be undirected. This argument is ignored for undirected
networks.}
}
\value{
For the conversion to sfnetwork: An object of class
\code{\link{sfnetwork}}. For the conversion from sfnetwork: a neighbor list,
which is a list with one element per node that holds the integer indices of
the nodes it is adjacent to.
}
\description{
Neighbor lists are sparse adjacency matrices in list format that specify for
each node to which other nodes it is adjacent. They occur for example in the
\code{\pkg{sf}} package as \code{\link[sf]{sgbp}} objects, and are also
frequently used in the \code{\pkg{spdep}} package.
}
