% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{nearest}
\alias{nearest}
\alias{nearest_nodes}
\alias{nearest_edges}
\title{Extract the nearest nodes or edges to given spatial features}
\usage{
nearest_nodes(x, y, focused = TRUE)

nearest_edges(x, y, focused = TRUE)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{y}{Spatial features as object of class \code{\link[sf]{sf}} or
\code{\link[sf]{sfc}}.}

\item{focused}{Should only features that are in focus be extracted? Defaults
to \code{TRUE}. See \code{\link[tidygraph]{focus}} for more information on
focused networks.}
}
\value{
An object of class \code{\link[sf]{sf}} with each row containing
the nearest node or edge to the corresponding spatial features in \code{y}.
}
\description{
Extract the nearest nodes or edges to given spatial features
}
\details{
To determine the nearest node or edge to each feature in \code{y}
the function \code{\link[sf]{st_nearest_feature}} is used. When extracting
nearest edges, spatially explicit edges are required, i.e. the edges table
should have a geometry column.
}
\examples{
library(sf, quietly = TRUE)

net = as_sfnetwork(roxel)
pts = st_sample(st_bbox(roxel), 6)

nodes = nearest_nodes(net, pts)
edges = nearest_edges(net, pts)

oldpar = par(no.readonly = TRUE)
par(mar = c(1,1,1,1), mfrow = c(1,2))

plot(net, main = "Nearest nodes")
plot(pts, cex = 2, col = "red", pch = 20, add = TRUE)
plot(st_geometry(nodes), cex = 2, col = "orange", pch = 20, add = TRUE)

plot(net, main = "Nearest edges")
plot(pts, cex = 2, col = "red", pch = 20, add = TRUE)
plot(st_geometry(edges), lwd = 2, col = "orange", pch = 20, add = TRUE)

par(oldpar)

}
