% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{nearest_ids}
\alias{nearest_ids}
\alias{nearest_node_ids}
\alias{nearest_edge_ids}
\title{Extract the indices of nearest nodes or edges to given spatial features}
\usage{
nearest_node_ids(x, y, focused = TRUE)

nearest_edge_ids(x, y, focused = TRUE)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{y}{Spatial features as object of class \code{\link[sf]{sf}} or
\code{\link[sf]{sfc}}.}

\item{focused}{Should only the indices of features that are in focus be
extracted? Defaults to \code{TRUE}. See \code{\link[tidygraph]{focus}} for
more information on focused networks.}
}
\value{
An integer vector with each element containing the index of the
nearest node or edge to the corresponding spatial features in \code{y}.
}
\description{
Extract the indices of nearest nodes or edges to given spatial features
}
\details{
To determine the nearest node or edge to each feature in \code{y}
the function \code{\link[sf]{st_nearest_feature}} is used. When extracting
nearest edges, spatially explicit edges are required, i.e. the edges table
should have a geometry column.
}
\examples{
library(sf, quietly = TRUE)

net = as_sfnetwork(roxel)
pts = st_sample(st_bbox(roxel))

nearest_node_ids(net, pts)
nearest_edge_ids(net, pts)

}
