% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{play_spatial}
\alias{play_spatial}
\alias{play_geometric}
\title{Create random spatial networks}
\usage{
play_geometric(
  n,
  radius,
  bounds = NULL,
  edges_as_lines = TRUE,
  compute_length = FALSE,
  ...
)
}
\arguments{
\item{n}{The number of nodes to be sampled.}

\item{radius}{The radius within which nodes will be connected by an edge.}

\item{bounds}{The spatial features within which the nodes should be sampled
as object of class \code{\link[sf]{sf}}, \code{\link[sf]{sfc}},
\code{\link[sf:st]{sfg}} or \code{\link[sf:st_bbox]{bbox}}. If set to
\code{NULL}, nodes will be sampled within a unit square.}

\item{edges_as_lines}{Should the created edges be spatially explicit, i.e.
have \code{LINESTRING} geometries stored in a geometry list column? Defaults
to \code{TRUE}.}

\item{compute_length}{Should the geographic length of the edges be stored in
a column named \code{length}? Uses \code{\link[sf]{st_length}} to compute
the length of the edge geometries when edges are spatially explicit, and
\code{\link[sf]{st_distance}} to compute the distance between boundary nodes
when edges are spatially implicit. Please note that the values in this
column are \strong{not} automatically recognized as edge weights. This needs
to be specified explicitly when calling a function that uses edge weights.
Defaults to \code{FALSE}.}

\item{...}{Additional arguments passed on to \code{\link[sf]{st_sample}}.
Ignored if \code{bounds = NULL}.}
}
\description{
Random spatial networks are created by randomly sampling nodes within a
given area, and connecting them by edges according to a specified method.
}
\section{Functions}{
\itemize{
\item \code{play_geometric()}: Creates a random geometric graph. Two nodes will be
connected by an edge if the distance between them is within the given radius.
If nodes are sampled on a unit square (i.e. when \code{bounds = NULL}) this
radius is unitless. If bounds are given as a spatial feature, the radius is
assumed to be in meters for geographic coordinates, and in the units of the
coordinate reference system for projected coordinates. Alternatively, units
can also be specified explicitly by providing a \code{\link[units]{units}}
object.

}}
\examples{
library(sf, quietly = TRUE)

oldpar = par(no.readonly = TRUE)
par(mar = c(1,1,1,1))

# Sample 10 nodes on a unit square
# Connect nodes by an edge if they are within 0.25 distance from each other
net = play_geometric(10, 0.25)
net

plot(net)

# Sample 10 nodes within a spatial bounding box
# Connect nodes by an edge if they are within 1 km from each other
net = play_geometric(10, units::set_units(1, "km"), bounds = st_bbox(roxel))
net

plot(net)

par(oldpar)

}
