% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.sfn_network}
\alias{plot}
\alias{plot.sfn_route}
\title{Plot a spatial network}
\usage{
\method{plot}{sfn_network}(x, y, ..., col_edges = "black", lwd = 1,
  col_nodes = "black", pch = 16, cex = 0.5, attribute = NULL,
  pal = NULL, add = FALSE)

\method{plot}{sfn_route}(x, y, ..., col_edges = "black", lwd = 1,
  col_nodes = "black", pch = 16, cex = 0.5, network = NULL,
  attribute = NULL, pal = NULL, add = FALSE)
}
\arguments{
\item{x}{object of class \code{sfn_network} or \code{sfn_route}}

\item{y}{ignored}

\item{...}{further specifications, passed on to \link{plot_sf}}

\item{col_edges}{color of the edges}

\item{lwd}{linewidth of the edges}

\item{col_nodes}{color of the nodes}

\item{pch}{plotting symbol of the nodes}

\item{cex}{symbol size of the nodes}

\item{attribute}{name of a column in the edges element of the input object, that should be used to color the edges;
by default set to \code{NULL};
if an attribute column is specified, only the edges will be plotted, and col_edges will be ignored}

\item{pal}{the color palette that is be used to color the edges by attribute;
if an attribute column is given but pal is omitted, \link{sf.colors} is used}

\item{add}{logical; if \code{TRUE}, plot will be added on top of the current plot;
by default set to \code{FALSE}}

\item{network}{object of class \code{sfn_network};
when plotting an object of class \code{sfn_route}, this parameter defines the network that should be plotted on the background;
by default set to \code{NULL}, meaning no network is plotted on the background}
}
\description{
Plot an object of class \code{sfn_network} or \code{sfn_route}.
}
\details{
The plot methods for objects of class \code{sfn_network} and \code{sfn_route}
uses the plot method for \code{sf} objects to plot the edges and noted.
See \link{plot_sf} for a more detailed description.
}
