% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sfnetwork}
\alias{plot.sfnetwork}
\title{Plot sfnetwork object}
\usage{
\method{plot}{sfnetwork}(x, ...)
}
\arguments{
\item{x}{object of class sfnetwork}

\item{...}{further parameters passed to \code{\link[sf]{plot}}}
}
\description{
Plot spatially explicit graph components (nodes and edges) of an sfnetwork on a map
}
