% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{points_to_line}
\alias{points_to_line}
\title{Draw a line between two points}
\usage{
points_to_line(x, y)
}
\arguments{
\item{x}{A \code{POINT} geometry.}

\item{y}{A \code{POINT} geometry.}
}
\value{
A \code{LINESTRING} geometry.
}
\description{
Draw a line between two points
}
