% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{points_to_lines}
\alias{points_to_lines}
\title{Draw lines between points}
\usage{
points_to_lines(sources, targets)
}
\arguments{
\item{sources}{An object of class \code{\link[sf]{sf}} with \code{POINT}
geometries, representing the points where lines need to start from.}

\item{targets}{An object of class \code{\link[sf]{sf}} with \code{POINT}
geometries, representing the points where lines need to end at.}
}
\value{
An object of class \code{\link[sf]{sf}} with \code{LINESTRING}
geometries.
}
\description{
Draw lines between points
}
