% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_geometry}
\alias{replace_geometry}
\title{Replace a geometry list column with another geometry list column}
\usage{
replace_geometry(x, y)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}} \code{\link[sf]{sf}} or
\code{\link[sf]{sfc}}.}

\item{y}{An object of class \code{\link[sf]{sfc}}.}
}
\value{
An object of the same class as x, but with a replaced geometry list
column
}
\description{
Replace a geometry list column with another geometry list column
}
