% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2.R
\name{s2}
\alias{s2}
\alias{as_s2_geography.sfnetwork}
\title{s2 methods for sfnetworks}
\usage{
\method{as_s2_geography}{sfnetwork}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{...}{Arguments passed on the corresponding \code{s2} function.}
}
\description{
s2 methods for sfnetworks
}
