% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{same_crs}
\alias{same_crs}
\title{Check if the CRS of two objects are the same}
\usage{
same_crs(x, y)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}, \code{\link[sf]{sf}} or
\code{\link[sf]{sfc}}.}

\item{y}{An object of class \code{\link{sfnetwork}}, \code{\link[sf]{sf}} or
\code{\link[sf]{sfc}}.}
}
\value{
\code{TRUE} when the CRS of x and y are the same, \code{FALSE}
otherwise.
}
\description{
Check if the CRS of two objects are the same
}
