% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{same_endpoints}
\alias{same_endpoints}
\title{Check if two sf objects have the same LINESTRING endpoints}
\usage{
same_endpoints(x, y)
}
\arguments{
\item{x}{An object of class \code{\link[sf]{sf}} or \code{\link[sf]{sfc}}
with \code{LINESTRING} geometries.}

\item{y}{An object of class \code{\link[sf]{sf}} or \code{\link[sf]{sfc}}
with \code{LINESTRING} geometries.}
}
\value{
\code{TRUE} when the endpoints are the same, \code{FALSE} otherwise.
}
\description{
Check if two sf objects have the same LINESTRING endpoints
}
\details{
This is a pairwise check. Each row in x is compared to its
corresponding row in y. Hence, x and y should be of the same length.
}
