% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrs.R
\name{sf_attr}
\alias{sf_attr}
\title{Query sf attributes from the active element of an sfnetwork}
\usage{
sf_attr(x, name = NULL, active = NULL)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{name}{Name of the attribute to query. If \code{NULL}, then all sf 
attributes are returned in a list. Defaults to \code{NULL}.}

\item{active}{Either 'nodes' or 'edges'. If \code{NULL}, the currently 
active element of x will be used.}
}
\value{
A named list of attributes if \code{name} is \code{NULL},
otherwise the value of the attribute matched, or NULL if no exact match is 
found and no or more than one partial match is found.
}
\description{
Query sf attributes from the active element of an sfnetwork
}
\details{
sf attributes include \code{sf_column} (the name of the sf column)
and \code{agr} (the attribute-geometry-relationships).
}
