% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_asnetwork.R
\name{sfn_asnetwork}
\alias{sfn_asnetwork}
\alias{sfn_asnetwork}
\alias{sfn_asnetwork.sf}
\alias{sfn_asnetwork}
\alias{sfn_asnetwork.SpatialLinesDataFrame}
\alias{sfn_asnetwork}
\alias{sfn_asnetwork.osmdata_sf}
\alias{sfn_asnetwork}
\alias{sfn_asnetwork.sfn_route}
\title{Convert foreign object to sfn_network object}
\usage{
sfn_asnetwork(x, attributes)

\method{sfn_asnetwork}{sf}(x, attributes = TRUE)

\method{sfn_asnetwork}{SpatialLinesDataFrame}(x, attributes = TRUE)

\method{sfn_asnetwork}{osmdata_sf}(x, attributes = TRUE)

\method{sfn_asnetwork}{sfn_route}(x, attributes = TRUE)
}
\arguments{
\item{x}{object to be converted into an sfn_network object; must be either of class \code{sf}, \code{SpatialLinesDataFrame}, \code{osmdata} or \code{sfn_route}}

\item{attributes}{logical;
if \code{TRUE}, copy attribute columns from the input object to the output object of class \code{sfn_network};
if \code{FALSE}, only take the geometry from the input object, and drop the attribute columns;
by default set to \code{TRUE}}
}
\value{
Returns an object of class \code{sfn_network},
which is a list representing a spatial network,
containing an sf object with edges (linestring geometry) and an sf object with nodes (point geometry)
}
\description{
Converts a foreign object to an an object of class \code{sfn_network}
}
