% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_length.R
\name{sfn_length}
\alias{sfn_length}
\alias{sfn_length}
\alias{sfn_length.sfn_network}
\alias{sfn_length}
\alias{sfn_length.sfn_route}
\title{Total length of a spatial network}
\usage{
sfn_length(x)

\method{sfn_length}{sfn_network}(x)

\method{sfn_length}{sfn_route}(x)
}
\arguments{
\item{x}{object of class \code{sfn_network} or \code{sfn_route}}
}
\value{
Returns the total length in meters.
}
\description{
Computes the total length of the edges in either a spatial network of class \code{sfn_network},
or a specific route in such a spatial network, of class \code{sfn_route}.
}
