% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_network2graph.R
\name{sfn_network2graph}
\alias{sfn_network2graph}
\title{Convert sfn_network object to igraph object}
\usage{
sfn_network2graph(x, weights = "length", directed = FALSE)
}
\arguments{
\item{x}{object of class \code{sfn_network}}

\item{weights}{either a string specifying the name of a column in the edges element of the sfn_network object,
which values should be used as edge weights;
or a numeric vector of the same length as the edge element of the sfn_network object,
containing the values that should be used as edge weights;
by default set to 'length', which is the name of the column containing the length in meters of each edge}

\item{directed}{logical; defines whether or not to create a directed graph;
by default set to \code{FALSE}}
}
\value{
Returns an object of class \code{igraph}, which is a class for mathematical graphs in the package \link{igraph}.
}
\description{
Converts an object of class \code{sfn_network} to an object of class \code{igraph}.
}
\details{
The spatial locations of the edges and nodes of the sfn_network object will not be maintained in the returned igraph object.
The returned graph is simply a set of weighted connections between unlocated nodes.
}
