% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_shortestpath.R
\name{sfn_shortestpath}
\alias{sfn_shortestpath}
\title{Shortest path between two nodes in a sfn_network}
\usage{
sfn_shortestpath(x, start, end, weights = "length", directed = FALSE)
}
\arguments{
\item{x}{object of class \code{sfn_network}}

\item{start}{unique node ID of the node from which the shortest path should be calculated}

\item{end}{unique node ID of the node to which the shortest path should lead}

\item{weights}{either a string specifying the name of a column in the edges element of the sfn_network object,
which values should be used as edge weights;
or a numeric vector of the same length as the edge element of the sfn_network object,
containing the values that should be used as edge weights;
by default set to 'length', which is the name of the column containing the length in meters of each edge}

\item{directed}{logical; defines whether or not to use a directed graph for calculating the shortest path;
by default set to \code{FALSE}}
}
\value{
Returns an object of class \code{sfn_route},
which is a list representing a specific route in a spatial network,
containing an sf object with the course of the route (geometrycollection),
an sf object with the edges that the route is made up of (linestring geometry),
an sf object with the nodes that the route is made up of (point geometry),
and a value representing the total weight of the route.
}
\description{
Calculates the shortest path between two nodes in a spatial network of class \code{sfn_network}.
}
