% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_transform.R
\name{sfn_transform}
\alias{sfn_transform}
\alias{sfn_transform}
\alias{sfn_transform.sfn_network}
\alias{sfn_transform}
\alias{sfn_transform.sfn_route}
\title{Transform or convert coordinates of sfn_network object}
\usage{
sfn_transform(x, crs, ...)

\method{sfn_transform}{sfn_network}(x, crs, ...)

\method{sfn_transform}{sfn_route}(x, crs, ...)
}
\arguments{
\item{x}{object of class \code{sfn_network} or \code{sfn_route}}

\item{crs}{the target coordinate reference system; must be either an integer with the EPSG code, or a character with proj4string}

\item{...}{further specification, passed on to \link{st_transform}}
}
\description{
Uses the function \link{st_transform} from the package \link{sf} to transform or convert coordinates
of both the edges and nodes element of an object of class \code{sfn_network},
or of both the route, edges and nodes element of an object of class \code{sfn_route}.
}
\details{
For a detailed description, see \link{st_transform}.
}
