% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify_network}
\alias{simplify_network}
\title{Simplify a spatial network}
\usage{
simplify_network(
  x,
  remove_multiple = TRUE,
  remove_loops = TRUE,
  attribute_summary = "first",
  store_original_ids = FALSE,
  store_original_data = FALSE
)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{remove_multiple}{Should multiple edges be merged into one. Defaults
to \code{TRUE}.}

\item{remove_loops}{Should loop edges be removed. Defaults to \code{TRUE}.}

\item{attribute_summary}{How should the attributes of merged multiple
edges be summarized? There are several options, see
\code{\link[igraph]{igraph-attribute-combination}} for details.}

\item{store_original_ids}{For each group of merged multiple edges, should
the indices of the original edges be stored as an attribute of the new edge,
in a column named \code{.tidygraph_edge_index}? This is in line with the
design principles of \code{tidygraph}. Defaults to \code{FALSE}.}

\item{store_original_data}{For each group of merged multiple edges, should
the data of the original edges be stored as an attribute of the new edge, in
a column named \code{.orig_data}? This is in line with the design principles
of \code{tidygraph}. Defaults to \code{FALSE}.}
}
\value{
The simple network as object of class \code{\link{sfnetwork}}.
}
\description{
Construct a simple version of the network. A simple network is defined as a
network without loop edges and multiple edges. A loop edge is an edge that
starts and ends at the same node. Multiple edges are different edges between
the same node pair.
}
\note{
When merging groups of multiple edges into a single edge, the geometry
of the first edge in each group is preserved. The order of the edges can be
influenced by calling \code{\link[dplyr]{arrange}} before simplifying.
}
