% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{spatial_centrality}
\alias{spatial_centrality}
\alias{centrality_straightness}
\title{Compute spatial centrality measures}
\usage{
centrality_straightness(...)
}
\arguments{
\item{...}{Additional arguments passed on to other functions.}
}
\value{
A numeric vector of the same length as the number of nodes in the
network.
}
\description{
These functions are a collection of centrality measures that are specific
for spatial networks, and form a spatial extension to
\code{\link[tidygraph:centrality]{centrality measures}} in tidygraph.
}
\details{
Just as with all centrality functions in tidygraph, these functions
are meant to be called inside tidygraph verbs such as
\code{\link[tidygraph]{mutate}} or \code{\link[tidygraph]{filter}}, where
the network that is currently being worked on is known and thus not needed
as an argument to the function. If you want to use an algorithm outside of
the tidygraph framework you can use \code{\link[tidygraph]{with_graph}} to
set the context temporarily while the algorithm is being evaluated.
}
\section{Functions}{
\itemize{
\item \code{centrality_straightness()}: The straightness centrality of node i is the
average ratio of Euclidean distance and network distance between node i and
all other nodes in the network. \code{...} is forwarded to
\code{\link{st_network_distance}} to compute the network distance matrix.
Euclidean distances are computed using \code{\link[sf]{st_distance}}.

}}
\examples{
library(tidygraph, quietly = TRUE)

net = as_sfnetwork(roxel, directed = FALSE)

net |>
  activate(nodes) |>
  mutate(sc = centrality_straightness())

}
