% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphers.R
\name{spatial_morphers}
\alias{spatial_morphers}
\alias{to_spatial_coordinates}
\alias{to_spatial_dense_graph}
\alias{to_spatial_directed}
\alias{to_spatial_explicit_edges}
\alias{to_spatial_implicit_edges}
\alias{to_spatial_shortest_paths}
\alias{to_spatial_subgraph}
\title{Spatial morphers for sfnetwork objects}
\usage{
to_spatial_coordinates(graph)

to_spatial_dense_graph(graph)

to_spatial_directed(graph)

to_spatial_explicit_edges(graph)

to_spatial_implicit_edges(graph)

to_spatial_shortest_paths(graph, ...)

to_spatial_subgraph(graph, ..., subset_by = NULL)
}
\arguments{
\item{graph}{An object of class \code{\link{sfnetwork}}.}

\item{...}{Arguments to be passed on to other function. See the description
of each morpher for details.}

\item{subset_by}{Whether to create subgraphs based on nodes or edges.}
}
\value{
A list of \code{\link{sfnetwork}} objects.
}
\description{
These functions are meant to be passed into [morph()] to create a temporary
alternate representation of the input graph. They are thus not meant to be
called directly. See below for detail of each morpher. For a more detailed
explanations on how to use morphing, see \code{\link[tidygraph]{morph}}.
}
\section{Functions}{
\itemize{
\item \code{to_spatial_coordinates}: Store the spatial coordinates of the nodes in 
separate coordinate columns (e.g. "X" and "Y"), instead of an 
\code{\link[sf]{sfc}} geometry list column. If edges are spatially
explicit, the edge geometries are dropped.

\item \code{to_spatial_dense_graph}: Reconstruct the network by using all edge 
linestring points as nodes, instead of only the endpoints.

\item \code{to_spatial_directed}: Make a graph directed in the direction given by 
the linestring geometries of the edges.

\item \code{to_spatial_explicit_edges}: Draw linestring geometries for spatially
implicit edges.

\item \code{to_spatial_implicit_edges}: Remove linestring geometries of spatially 
explicit edges.

\item \code{to_spatial_shortest_paths}: Limit a graph to those nodes and edges that are
part of the shortest path between two nodes. If multiple \code{to} nodes are 
given, multiple shortest paths are returned. \code{...} is evaluated in the 
same manner as \code{\link{st_shortest_paths}}, except that the \code{output},
\code{predecessors} and \code{inbound.edges} arguments are ignored. When 
unmorphing only the first instance of both the node and edge data will be 
used, as the the same node and/or edge can be present in multiple paths.

\item \code{to_spatial_subgraph}: Limit a graph to a single spatial subset. 
\code{...} is evaluated in the same manner as \code{\link{st_filter}}.
}}

