% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{st_duplicated}
\alias{st_duplicated}
\title{Determine duplicated geometries}
\usage{
st_duplicated(x)
}
\arguments{
\item{x}{An object of class \code{\link[sf]{sf}} or \code{\link[sf]{sfc}}.}
}
\value{
A logical vector specifying for each feature in \code{x} if its
geometry is equal to a previous feature in \code{x}.
}
\description{
Determine duplicated geometries
}
\examples{
library(sf, quietly = TRUE)

p1 = st_sfc(st_point(c(1, 1)))
p2 = st_sfc(st_point(c(0, 0)))
p3 = st_sfc(st_point(c(1, 0)))

st_duplicated(c(p1, p2, p2, p3, p1))

}
