% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{st_is_all}
\alias{st_is_all}
\title{Check if the geometries of an sf object are all of a specific type}
\usage{
st_is_all(x, type)
}
\arguments{
\item{x}{An object of class \code{\link[sf]{sf}}.}

\item{type}{The geometry type to check for, as a string.}
}
\value{
\code{TRUE} when all geometries are of the given type, \code{FALSE}
otherwise.
}
\description{
Check if the geometries of an sf object are all of a specific type
}
