% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{st_match}
\alias{st_match}
\title{Geometry matching}
\usage{
st_match(x)
}
\arguments{
\item{x}{An object of class \code{\link[sf]{sf}} or \code{\link[sf]{sfc}}.}
}
\value{
A numeric vector giving for each feature in \code{x} the position of
the first feature in \code{x} that has an equal geometry.
}
\description{
Geometry matching
}
\examples{
library(sf, quietly = TRUE)

p1 = st_sfc(st_point(c(1, 1)))
p2 = st_sfc(st_point(c(0, 0)))
p3 = st_sfc(st_point(c(1, 0)))

st_match(c(p1, p2, p2, p3, p1))

}
