% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faces.R
\name{st_network_faces}
\alias{st_network_faces}
\title{Extract the faces of a spatial network}
\usage{
st_network_faces(x, boundary = NULL)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{boundary}{The boundary used for the outer face, as an object of class
\code{\link[sf]{sf}} or \code{\link[sf]{sfc}} containing a single
\code{POLYGON} geometry. Note that this boundary should always be larger
than the bounding box of the network. If \code{NULL} (the default) the
network bounding box extended by 0.1 times its diameter is used.}
}
\value{
An object of class \code{\link[sf]{sfc}} with \code{POLYGON}
geometries, in which each feature represents one face of the network.
}
\description{
The faces of a spatial network are the areas bounded by edges, without any
other edge crossing it. A special face is the outer face, which is the area
not bounded by any set of edges.
}
\examples{
library(sf, quietly = TRUE)

oldpar = par(no.readonly = TRUE)
par(mar = c(1,1,1,1))

pts = st_transform(mozart, 3035)
net = as_sfnetwork(pts, "delaunay")

faces = st_network_faces(net)

plot(faces, col = sf.colors(length(faces), categorical = TRUE))
plot(net, add = TRUE)

par(oldpar)

}
