% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{st_network_travel}
\alias{st_network_travel}
\title{Compute route optimization algorithms}
\usage{
st_network_travel(
  x,
  pois,
  weights = edge_length(),
  algorithm = "tsp",
  return_paths = TRUE,
  use_names = TRUE,
  return_cost = TRUE,
  return_geometry = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{pois}{Locations that the travelling salesman will visit. Evaluated by
\code{\link{evaluate_node_query}}.}

\item{weights}{The edge weights to be used in the shortest path calculation.
Evaluated by \code{\link{evaluate_edge_spec}}. The default is
\code{\link{edge_length}}, which computes the geographic lengths of the
edges.}

\item{return_paths}{Should the shortest paths between `pois` be computed?
Defaults to `TRUE`. If `FALSE`, a vector with indices in the visiting order
is returned.}

\item{use_names}{If a column named \code{name} is present in the nodes
table, should these names be used to encode the nodes in a path, instead of
the node indices? Defaults to \code{TRUE}. Ignored when the nodes table does
not have a column named \code{name}.}

\item{return_cost}{Should the total cost of each path be computed? Defaults
to \code{TRUE}. Ignored if \code{type = 'all_simple'}.}

\item{return_geometry}{Should a linestring geometry be constructed for each
path? Defaults to \code{TRUE}. The geometries are constructed by calling
\code{\link[sf]{st_line_merge}} on the linestring geometries of the edges in
the path. Ignored if \code{type = 'all_simple'} and for networks with
spatially implicit edges.}

\item{...}{Additional arguments passed on to the `TSP::solve_tsp()` function.}
}
\value{
An object of class \code{\link[tibble]{tbl_df}} or
\code{\link[sf]{sf}} with one row per path, or a vector with ordered indices
for `pois`.
}
\description{
The travelling salesman problem is currently implemented
}
