% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap.R
\name{st_snap_to_network}
\alias{st_snap_to_network}
\title{Update geometries to their nearest point on the network}
\usage{
st_snap_to_network(x, graph, method = "nearest_node")
}
\arguments{
\item{x}{The spatial features to be snapped, either as object of class
\code{\link[sf]{sf}} or \code{\link[sf]{sfc}}.}

\item{graph}{The network to be snapped to as object object of class
\code{\link{sfnetwork}}.}

\item{method}{The snapping technique to be used. See details. Defaults to
\code{'nearest_node'}.}
}
\value{
An object of class \code{\link[sf]{sfc}}, containing the snapped
geometries of x.
}
\description{
Implementation of snapping techniques that find the nearest point on a
spatial network to a given set of input geometries.
}
\details{
Implemented snapping techniques are.
\itemize{
\item \code{'nearest_node'}: Find the nearest node to the given input geometry.
\item \code{'nearest_point_on_edge'}: Finds the nearest point on the nearest
edge to the given input geometry, no matter if this point is a node in the
network or a line vertex in the linestring geometry of the edge. This
technique requires spatially explicit edges.
}
}
