% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygraph.R
\name{tidygraph}
\alias{tidygraph}
\alias{as_tbl_graph.sfnetwork}
\alias{as_tibble.sfnetwork}
\alias{morph.sfnetwork}
\title{tidygraph methods for sfnetwork objects}
\usage{
\method{as_tbl_graph}{sfnetwork}(x, ...)

\method{as_tibble}{sfnetwork}(x, active = NULL, spatial = TRUE, ...)

\method{morph}{sfnetwork}(.data, .f, ...)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{...}{Arguments passed on the corresponding \code{tidygraph} function.}

\item{active}{Which network element (i.e. nodes or edges) to activate before
extracting. If \code{NULL}, it will be set to the current active element of
the given network. Defaults to \code{NULL}.}

\item{spatial}{Should te extracted tibble be a 'spatial tibble', i.e. an
object of class \code{c('sf', 'tbl_df')}, if it contains a geometry list
column. Defaults to \code{TRUE}.}

\item{.data}{An object of class \code{\link{sfnetwork}}.}

\item{.f}{See \code{\link[tidygraph]{morph}}.}
}
\description{
\code{\link[tidygraph:tidygraph-package]{tidygraph}} methods for 
\code{\link{sfnetwork}} objects. Since \code{\link{sfnetwork}} objects
subclass \code{\link[tidygraph]{tbl_graph}} objects, most tidygraph
functions work automatically. However, some of them need a special method,
mostly as a result of the presence of the geometry list column in
\code{\link{sfnetwork}} objects. Use these methods without the .sfnetwork 
suffix and after loading the tidygraph package.
}
\details{
See the \code{\link[tidygraph:tidygraph-package]{tidygraph}} 
documentation.
}
