% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygraph.R
\name{tidygraph_methods}
\alias{tidygraph_methods}
\alias{as_tbl_graph.sfnetwork}
\alias{reroute.sfnetwork}
\alias{morph.sfnetwork}
\alias{unmorph.morphed_sfnetwork}
\title{tidygraph methods for sfnetworks}
\usage{
\method{as_tbl_graph}{sfnetwork}(x, ...)

\method{reroute}{sfnetwork}(.data, ...)

\method{morph}{sfnetwork}(.data, ...)

\method{unmorph}{morphed_sfnetwork}(.data, ...)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{...}{Arguments passed on the corresponding \code{tidygraph} function.}

\item{.data}{An object of class \code{\link{sfnetwork}}.}
}
\value{
The method for \code{\link[tidygraph]{as_tbl_graph}} returns an
object of class \code{\link[tidygraph]{tbl_graph}}. The method for
\code{\link[tidygraph]{morph}} returns a \code{morphed_sfnetwork} if the
morphed network is still spatial, and a \code{morphed_tbl_graph} otherwise.
All other methods return an object of class \code{\link{sfnetwork}}.
}
\description{
Normally tidygraph functions should work out of the box on
\code{\link{sfnetwork}} objects, but in some cases special treatment is
needed especially for the geometry column, requiring a specific method.
}
\details{
See the \code{\link[tidygraph]{tidygraph}} documentation. The
following methods have a special behavior:

\itemize{
  \item \code{reroute}: To preserve the valid spatial network structure,
  this method will replace the boundaries of edge geometries by the location
  of the node those edges are rerouted to or from. Note that when the goal
  is to reverse edges in a spatial network, reroute will not simply reverse
  the edge geometries. In that case it is recommended to use the sfnetwork
  method for \code{\link[sf]{st_reverse}} instead.
  \item \code{morph}: This method checks if the morphed network still has
  spatially embedded nodes. In that case a \code{morphed_sfnetwork} is
  returned. If not, a \code{morphed_tbl_graph} is returned instead.
  \item \code{unmorph}: This method makes sure the geometry list column is
  correctly handled during the unmorphing process.
}
}
