% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_network}
\alias{validate_network}
\title{Validate the structure of a sfnetwork}
\usage{
validate_network(x, message = TRUE)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{message}{Should messages be printed during validation? Defaults to
\code{TRUE}.}
}
\value{
Nothing when the network is valid. Otherwise, an error is thrown.
}
\description{
Validate the structure of a sfnetwork
}
\details{
A valid sfnetwork structure means that all nodes have \code{POINT}
geometries, and - when edges are spatially explicit - all edges have
\code{LINESTRING} geometries, nodes and edges have the same coordinate
reference system and the same coordinate precision, and coordinates of
edge boundaries match coordinates of their corresponding nodes.
}
