% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linnet.R
\name{as.linnet}
\alias{as.linnet}
\alias{as.linnet.sfnetwork}
\title{Convert a sfnetwork into a linnet}
\usage{
\method{as.linnet}{sfnetwork}(X, ...)
}
\arguments{
\item{X}{An object of class \code{\link{sfnetwork}} with a projected CRS.}

\item{...}{Arguments passed to \code{\link[spatstat]{linnet}}.}
}
\value{
An object of class \code{\link[spatstat]{linnet}}.
}
\description{
A method to convert an object of class \code{\link{sfnetwork}} into
\code{\link[spatstat]{linnet}} format and enhance the interoperability
between \code{sfnetworks} and \code{spatstat}. Use this method without the
.sfnetwork suffix and after loading the \code{spatstat} package.
}
\seealso{
\code{\link{as_sfnetwork}} to convert objects of class
\code{\link[spatstat]{linnet}} into objects of class
\code{\link{sfnetwork}}.
}
