% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{as_sfnetwork}
\alias{as_sfnetwork}
\alias{as_sfnetwork.default}
\alias{as_sfnetwork.sf}
\alias{as_sfnetwork.sfc}
\alias{as_sfnetwork.linnet}
\alias{as_sfnetwork.psp}
\alias{as_sfnetwork.sfNetwork}
\alias{as_sfnetwork.tbl_graph}
\title{Convert a foreign object to a sfnetwork}
\usage{
as_sfnetwork(x, ...)

\method{as_sfnetwork}{default}(x, ...)

\method{as_sfnetwork}{sf}(x, ...)

\method{as_sfnetwork}{sfc}(x, ...)

\method{as_sfnetwork}{linnet}(x, ...)

\method{as_sfnetwork}{psp}(x, ...)

\method{as_sfnetwork}{sfNetwork}(x, ...)

\method{as_sfnetwork}{tbl_graph}(x, ...)
}
\arguments{
\item{x}{Object to be converted into a \code{\link{sfnetwork}}.}

\item{...}{Additional arguments passed on to the \code{\link{sfnetwork}}
construction function, unless specified otherwise.}
}
\value{
An object of class \code{\link{sfnetwork}}.
}
\description{
Convert a given object into an object of class \code{\link{sfnetwork}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{as_sfnetwork(default)}: By default, the provided object is first converted
into a \code{\link[tidygraph]{tbl_graph}} using
\code{\link[tidygraph]{as_tbl_graph}}. Further conversion into an
\code{\link{sfnetwork}} will work as long as the nodes can be converted to
an \code{\link[sf]{sf}} object through \code{\link[sf]{st_as_sf}}. Arguments
to \code{\link[sf]{st_as_sf}} can be provided as additional arguments and
will be forwarded to \code{\link[sf]{st_as_sf}} through the
code{\link{sfnetwork}} construction function.

\item \code{as_sfnetwork(sf)}: Convert spatial features of class
\code{\link[sf]{sf}} directly into a \code{\link{sfnetwork}}.
Supported geometry types are either \code{LINESTRING} or \code{POINT}. In
the first case, the lines become the edges in the network, and nodes are
placed at their boundaries. Additional arguments are forwarded to
\code{\link{create_from_spatial_lines}}. In the latter case, the points
become the nodes in the network, and are connected by edges according to a
specified method. Additional arguments are forwarded to
\code{\link{create_from_spatial_points}}.

\item \code{as_sfnetwork(sfc)}: Convert spatial geometries of class
\code{\link[sf]{sfc}} directly into a \code{\link{sfnetwork}}.
Supported geometry types are either \code{LINESTRING} or \code{POINT}. In
the first case, the lines become the edges in the network, and nodes are
placed at their boundaries. Additional arguments are forwarded to
\code{\link{create_from_spatial_lines}}. In the latter case, the points
become the nodes in the network, and are connected by edges according to a
specified method. Additional arguments are forwarded to
\code{\link{create_from_spatial_points}}.

\item \code{as_sfnetwork(linnet)}: Convert spatial linear networks of class
\code{\link[spatstat.linnet]{linnet}} directly into an
\code{\link{sfnetwork}}. This requires the
\code{\link[spatstat.geom:spatstat.geom-package]{spatstat.geom}} package
to be installed.

\item \code{as_sfnetwork(psp)}: Convert spatial line segments of class
\code{\link[spatstat.geom]{psp}} directly into a \code{\link{sfnetwork}}.
The lines become the edges in the network, and nodes are placed at their
boundary points.

\item \code{as_sfnetwork(sfNetwork)}: Convert spatial networks of class
\code{\link[stplanr:sfNetwork-class]{sfNetwork}} directly into a
\code{\link{sfnetwork}}. This will extract the edges as an
\code{\link[sf]{sf}} object and re-create the network structure. The
directness of the original network is preserved unless specified otherwise
through the \code{directed} argument.

\item \code{as_sfnetwork(tbl_graph)}: Convert graph objects of class
\code{\link[tidygraph]{tbl_graph}} directly into a \code{\link{sfnetwork}}.
This will work if at least the nodes can be converted to an
\code{\link[sf]{sf}} object through \code{\link[sf]{st_as_sf}}. The
directness of the original graph is preserved unless specified otherwise
through the \code{directed} argument.

}}
\examples{
# From an sf object with LINESTRING geometries.
library(sf, quietly = TRUE)

oldpar = par(no.readonly = TRUE)
par(mar = c(1,1,1,1), mfrow = c(1,2))

as_sfnetwork(roxel)

plot(st_geometry(roxel))
plot(as_sfnetwork(roxel))

# From an sf object with POINT geometries.
# For more examples see ?create_from_spatial_points.
as_sfnetwork(mozart)

plot(st_geometry(mozart))
plot(as_sfnetwork(mozart))

par(oldpar)

# From a linnet object.
if (require(spatstat.geom, quietly = TRUE)) {
  as_sfnetwork(simplenet)
}

# From a psp object.
if (require(spatstat.geom, quietly = TRUE)) {
  set.seed(42)
  test_psp = psp(runif(10), runif(10), runif(10), runif(10), window=owin())
  as_sfnetwork(test_psp)
}

# From a tbl_graph with coordinate columns.
library(tidygraph, quietly = TRUE)

nodes = data.frame(lat = c(7, 7, 8), lon = c(51, 52, 52))
edges = data.frame(from = c(1, 1, 3), to = c(2, 3, 2))
tbl_net = tbl_graph(nodes, edges)
as_sfnetwork(tbl_net, coords = c("lon", "lat"), crs = 4326)

}
