% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfnetwork.R
\name{as_sfnetwork}
\alias{as_sfnetwork}
\alias{as_sfnetwork.default}
\alias{as_sfnetwork.linnet}
\alias{as_sfnetwork.psp}
\alias{as_sfnetwork.sf}
\alias{as_sfnetwork.sfc}
\alias{as_sfnetwork.sfNetwork}
\alias{as_sfnetwork.sfnetwork}
\alias{as_sfnetwork.tbl_graph}
\title{Convert a foreign object to a sfnetwork}
\usage{
as_sfnetwork(x, ...)

\method{as_sfnetwork}{default}(x, ...)

\method{as_sfnetwork}{linnet}(x, ...)

\method{as_sfnetwork}{psp}(x, ...)

\method{as_sfnetwork}{sf}(x, ...)

\method{as_sfnetwork}{sfc}(x, ...)

\method{as_sfnetwork}{sfNetwork}(x, ...)

\method{as_sfnetwork}{sfnetwork}(x, ...)

\method{as_sfnetwork}{tbl_graph}(x, ...)
}
\arguments{
\item{x}{Object to be converted into an \code{\link{sfnetwork}}.}

\item{...}{Arguments passed on to the \code{\link{sfnetwork}} construction
function.}
}
\value{
An object of class \code{\link{sfnetwork}}.
}
\description{
Convert a given object into an object of class \code{\link{sfnetwork}}.
If an object can be read by \code{\link[tidygraph]{as_tbl_graph}} and the
nodes can be read by \code{\link[sf]{st_as_sf}}, it is automatically
supported.
}
\examples{
# From a linnet object.
if (require(spatstat)) as_sfnetwork(simplenet)

# From a psp object.
if (require(spatstat)) {
  set.seed(42)
  test_psp = psp(runif(10), runif(10), runif(10), runif(10), window=owin())
  as_sfnetwork(test_psp)
}

# From an sf object with POINT geometries.
library(sf)
p1 = st_point(c(7, 51))
p2 = st_point(c(7, 52))
p3 = st_point(c(8, 52))
points = st_as_sf(st_sfc(p1, p2, p3, crs = 4326))
as_sfnetwork(points)

# From an sf object with LINESTRING geometries.
library(sf)
e1 = st_cast(st_union(p1,p2), "LINESTRING")
e2 = st_cast(st_union(p1,p3), "LINESTRING")
e3 = st_cast(st_union(p2,p3), "LINESTRING")
lines = st_as_sf(st_sfc(e1, e2, e3, crs = 4326))
as_sfnetwork(lines)

}
