% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.sfnetwork}
\alias{autoplot.sfnetwork}
\title{Autoplot sfnetwork object}
\usage{
\method{autoplot}{sfnetwork}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{sfnetwork}}.}

\item{...}{ignored}
}
\description{
Autoplot an object of class \code{\link{sfnetwork}} as a ggplot object.
}
\details{
This is a basic ggploting functionality, for a quick
assessment of the network but on a \code{ggplot2} format.
}
\examples{
library(ggplot2)
library(sf)
net = as_sfnetwork(roxel)

# Quick overview of the network in `ggplot2` style
autoplot(net)

# Other `ggplot2` elements can be added
points = net \%>\%
  st_bbox() \%>\%
  st_as_sfc() \%>\%
  st_sample(10, type = 'random') \%>\%
  st_set_crs(4326) \%>\%
  st_cast('POINT')

autoplot(net) +
   # The theme can be customized
   theme_minimal() +
   # Labels can be added
   labs(title = 'Nice ggplot') +
   # And extra `geom_sf` layers can be included
   geom_sf(data = points, color = 'red', size = 2)
}
