% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capture_plain}
\alias{capture_plain}
\title{Subset and modify the result of \code{capture.output} to pass onto print
method}
\usage{
capture_plain(p, start, end, prefix = "", sep = "\\n")
}
\arguments{
\item{p}{The result of \code{capture.output}.}

\item{start}{From which line to start subsetting.}

\item{end}{From which line to finish subsetting.}

\item{prefix}{A string that should be passed before printing each line.}

\item{sep}{How to separate each line for printing.}
}
\value{
A modified subset of the original captured output.
}
\description{
Subset and modify the result of \code{capture.output} to pass onto print
method
}
