% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capture_subtle}
\alias{capture_subtle}
\title{Capture output from console to pass onto print method with subtle style}
\usage{
capture_subtle(p, start, end, prefix = "", sep = "\\n")
}
\arguments{
\item{p}{The result of a print method on any class object.}

\item{start}{from which line to start capturing output.}

\item{end}{from which line to finish capturing output.}

\item{prefix}{A string that should be passed before printing each line}

\item{sep}{How to separate each line for printing}
}
\value{
A subset of the original printed object passed to the console
with subtle style
}
\description{
Capture output from console to pass onto print method with subtle style
}
