% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_from_spatial_lines}
\alias{create_from_spatial_lines}
\title{Create a spatial network from linestring geometries}
\usage{
create_from_spatial_lines(x, directed = TRUE, compute_length = FALSE)
}
\arguments{
\item{x}{An object of class \code{\link[sf]{sf}} or \code{\link[sf]{sfc}}
with \code{LINESTRING} geometries.}

\item{directed}{Should the constructed network be directed? Defaults to
\code{TRUE}.}

\item{compute_length}{Should the geographic length of the edges be stored in
a column named \code{length}? Uses \code{\link[sf]{st_length}} to compute
the length of the edge geometries. If there is already a column named
\code{length}, it will be overwritten. Please note that the values in this
column are \strong{not} automatically recognized as edge weights. This needs
to be specified explicitly when calling a function that uses edge weights.
Defaults to \code{FALSE}.}
}
\value{
An object of class \code{\link{sfnetwork}}.
}
\description{
Create a spatial network from linestring geometries
}
\details{
It is assumed that the given linestring geometries form the edges
in the network. Nodes are created at the line boundaries. Shared boundaries
between multiple linestrings become the same node.
}
\note{
By default sfnetworks rounds coordinates to 12 decimal places to
determine spatial equality. You can influence this behavior by explicitly
setting the precision of the linestrings using
\code{\link[sf]{st_set_precision}}.
}
\examples{
library(sf, quietly = TRUE)

as_sfnetwork(roxel)

oldpar = par(no.readonly = TRUE)
par(mar = c(1,1,1,1), mfrow = c(1,2))

plot(st_geometry(roxel))
plot(as_sfnetwork(roxel))

par(oldpar)

}
