% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_from_spatial_lines}
\alias{create_from_spatial_lines}
\title{Create a spatial network from linestring geometries}
\usage{
create_from_spatial_lines(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link[sf]{sf}} or \code{\link[sf]{sfc}}
with \code{LINESTRING} geometries.}
}
\value{
An object of class \code{\link{sfnetwork}}.
}
\description{
Create a spatial network from linestring geometries
}
\details{
It is assumed that the given lines geometries form the edges in the
network. Nodes are created at the boundary points of the edges. Boundary
points at equal locations become the same node.
}
