% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_from_spatial_points}
\alias{create_from_spatial_points}
\title{Create a spatial network from point geometries}
\usage{
create_from_spatial_points(x, method = "sequence", ...)
}
\arguments{
\item{x}{An object of class \code{\link[sf]{sf}} or \code{\link[sf]{sfc}}
with \code{POINT} geometries.}

\item{method}{The method used to connect the given point geometries to each
other.}
}
\value{
An object of class \code{\link{sfnetwork}}.
}
\description{
Create a spatial network from point geometries
}
\details{
It is assumed that the given points form the nodes in the network.
How those nodes are connected by edges depends on the selected method.
}
