% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{ids}
\alias{ids}
\alias{node_ids}
\alias{edge_ids}
\title{Extract the node or edge indices from a spatial network}
\usage{
node_ids(x, focused = TRUE)

edge_ids(x, focused = TRUE)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{focused}{Should only the indices of features that are in focus be
extracted? Defaults to \code{TRUE}. See \code{\link[tidygraph]{focus}} for
more information on focused networks.}
}
\value{
An vector of integers.
}
\description{
Extract the node or edge indices from a spatial network
}
\details{
The indices in these objects are always integers that correspond to
rownumbers in respectively the nodes or edges table.
}
\examples{
net = as_sfnetwork(roxel[1:10, ])
node_ids(net)
edge_ids(net)

}
