% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ids}
\alias{ids}
\alias{node_ids}
\alias{edge_ids}
\title{Extract the indices of nodes or edges from a network}
\usage{
node_ids(x)

edge_ids(x)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}, or any other network
object inheriting from \code{\link[igraph]{igraph}}.}
}
\value{
An vector of integers.
}
\description{
Extract the indices of nodes or edges from a network
}
\details{
The indices in these objects are always integers that correspond to
rownumbers in respectively the nodes or edges table.
}
\examples{
net = as_sfnetwork(roxel[1:10, ])
node_ids(net)
edge_ids(net)

}
