% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfnetwork.R
\name{is.sfnetwork}
\alias{is.sfnetwork}
\title{Check if an object is a sfnetwork}
\usage{
is.sfnetwork(x)
}
\arguments{
\item{x}{Object to be checked.}
}
\value{
\code{TRUE} if the given object is an object of class
\code{\link{sfnetwork}}, \code{FALSE} otherwise.
}
\description{
Check if an object is a sfnetwork
}
\examples{
library(tidygraph, quietly = TRUE, warn.conflicts = FALSE)

net = as_sfnetwork(roxel)
is.sfnetwork(net)
is.sfnetwork(as_tbl_graph(net))

}
