% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.R
\name{make_edges_explicit}
\alias{make_edges_explicit}
\title{Construct edge geometries for spatially implicit networks}
\usage{
make_edges_explicit(x)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}
}
\value{
An object of class \code{\link{sfnetwork}} with spatially explicit
edges. If \code{x} was already spatially explicit it is returned unmodified.
}
\description{
This function turns spatially implicit networks into spatially explicit
networks by adding a geometry column to the edges data containing straight
lines between the source and target nodes.
}
