% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sfnetwork}
\alias{plot.sfnetwork}
\title{Plot sfnetwork object}
\usage{
\method{plot}{sfnetwork}(x, ...)
}
\arguments{
\item{x}{Object of class \code{\link{sfnetwork}}.}

\item{...}{Arguments passed on to \code{\link[sf]{plot}}}
}
\description{
Plot the geometries of an object of class \code{\link{sfnetwork}}.
}
