% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygraph.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{activate}
\alias{active}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tidygraph}{\code{\link[tidygraph]{activate}}, \code{\link[tidygraph]{active}}}
}}

