% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/tidygraph.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{activate}
\alias{active}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{tidygraph}{\code{\link[tidygraph:reexports]{\%>\%}}, \code{\link[tidygraph]{activate}}, \code{\link[tidygraph:activate]{active}}}
}}

