% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sf}
\alias{sf}
\alias{st_as_sf.sfnetwork}
\alias{st_bbox.sfnetwork}
\alias{st_crs.sfnetwork}
\alias{st_set_crs.sfnetwork}
\alias{st_geometry.sfnetwork}
\alias{st_make_grid.sfnetwork}
\alias{st_contains.sfnetwork}
\alias{st_contains_properly.sfnetwork}
\alias{st_covers.sfnetwork}
\alias{st_covered_by.sfnetwork}
\alias{st_crosses.sfnetwork}
\alias{st_disjoint.sfnetwork}
\alias{st_equals.sfnetwork}
\alias{st_equals_exact.sfnetwork}
\alias{st_intersects.sfnetwork}
\alias{st_is_within_distance.sfnetwork}
\alias{st_overlaps.sfnetwork}
\alias{st_touches.sfnetwork}
\alias{st_within.sfnetwork}
\title{sf methods for sfnetwork objects}
\usage{
\method{st_as_sf}{sfnetwork}(x, ...)

\method{st_bbox}{sfnetwork}(x, ...)

\method{st_crs}{sfnetwork}(x, ...)

st_set_crs.sfnetwork(x, ...)

\method{st_geometry}{sfnetwork}(x, ...)

st_make_grid.sfnetwork(x, ...)

st_contains.sfnetwork(x, y, ...)

st_contains_properly.sfnetwork(x, y, ...)

st_covers.sfnetwork(x, y, ...)

st_covered_by.sfnetwork(x, y, ...)

st_crosses.sfnetwork(x, y, ...)

st_disjoint.sfnetwork(x, y, ...)

st_equals.sfnetwork(x, y, ...)

st_equals_exact.sfnetwork(x, y, ...)

\method{st_intersects}{sfnetwork}(x, y, ...)

st_is_within_distance.sfnetwork(x, y, ...)

st_overlaps.sfnetwork(x, y, ...)

st_touches.sfnetwork(x, y, ...)

st_within.sfnetwork(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{...}{Arguments passed on the corresponding \code{sf} function.}

\item{y}{An object of class \link{sfnetwork} or class \code{\link[sf]{sf}}.}
}
\description{
Tidyverse methods for sf objects. Use these methods without the .sfnetwork
suffix and after loading the sf package.
}
