% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sf}
\alias{sf}
\alias{st_as_sf.sfnetwork}
\alias{st_geometry.sfnetwork}
\alias{st_geometry<-.sfnetwork}
\alias{st_bbox.sfnetwork}
\alias{st_coordinates.sfnetwork}
\alias{st_is.sfnetwork}
\alias{st_crs.sfnetwork}
\alias{st_crs<-.sfnetwork}
\alias{st_shift_longitude.sfnetwork}
\alias{st_transform.sfnetwork}
\alias{st_wrap_dateline.sfnetwork}
\alias{st_zm.sfnetwork}
\alias{st_m_range.sfnetwork}
\alias{st_z_range.sfnetwork}
\alias{st_agr.sfnetwork}
\alias{st_agr<-.sfnetwork}
\alias{st_intersects.sfnetwork}
\alias{st_reverse.sfnetwork}
\alias{st_simplify.sfnetwork}
\alias{st_join.sfnetwork}
\alias{st_crop.sfnetwork}
\alias{st_filter.sfnetwork}
\title{sf methods for sfnetworks}
\usage{
\method{st_as_sf}{sfnetwork}(x, active = NULL, ...)

\method{st_geometry}{sfnetwork}(x, active = NULL, ...)

\method{st_geometry}{sfnetwork}(x) <- value

\method{st_bbox}{sfnetwork}(x, ...)

\method{st_coordinates}{sfnetwork}(x, ...)

\method{st_is}{sfnetwork}(x, ...)

\method{st_crs}{sfnetwork}(x, ...)

\method{st_crs}{sfnetwork}(x) <- value

\method{st_shift_longitude}{sfnetwork}(x, ...)

\method{st_transform}{sfnetwork}(x, ...)

\method{st_wrap_dateline}{sfnetwork}(x, ...)

\method{st_zm}{sfnetwork}(x, ...)

\method{st_m_range}{sfnetwork}(x, ...)

\method{st_z_range}{sfnetwork}(x, ...)

\method{st_agr}{sfnetwork}(x, active = NULL, ...)

\method{st_agr}{sfnetwork}(x) <- value

\method{st_intersects}{sfnetwork}(x, y = x, ...)

\method{st_reverse}{sfnetwork}(x, ...)

\method{st_simplify}{sfnetwork}(x, ...)

\method{st_join}{sfnetwork}(x, y, ...)

\method{st_crop}{sfnetwork}(x, y, ...)

\method{st_filter}{sfnetwork}(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{active}{Which network element (i.e. nodes or edges) to activate before
extracting. If \code{NULL}, it will be set to the current active element of
the given network. Defaults to \code{NULL}.}

\item{...}{Arguments passed on the corresponding \code{sf} function.}

\item{value}{See \code{\link[sf]{st_crs}}, \code{\link[sf]{st_geometry}} or 
\code{\link[sf]{st_agr}}.}

\item{y}{An object of class \code{\link[sf]{sf}}, or directly convertible to 
it using \code{\link[sf]{st_as_sf}}. In some cases, it can also be an object 
of \code{\link[sf:st]{sfg}} or \code{\link[sf:st_bbox]{bbox}}. Always look 
at the documentation of the corresponding \code{sf} function for details.}
}
\description{
\code{\link[sf]{sf}} methods for \code{\link{sfnetwork}} objects. Use these 
methods without the .sfnetwork suffix and after loading the sf package.
}
\details{
See the \code{\link[sf]{sf}} documentation.
}
