% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrs.R
\name{sf_attr}
\alias{sf_attr}
\alias{sf_attr<-}
\title{Query sf attributes from the active element of an sfnetwork object}
\usage{
sf_attr(x, name = NULL, active = NULL)

sf_attr(x, name, active = NULL) <- value
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{name}{Name of the attribute to query. If \code{NULL}, then all sf 
attributes are returned in a list. Defaults to \code{NULL}.}

\item{active}{Which network element (i.e. nodes or edges) to activate before
extracting. If \code{NULL}, it will be set to the current active element of
the given network. Defaults to \code{NULL}.}

\item{value}{The new value of the attribute, or \code{NULL} to remove the 
attribute.}
}
\value{
For the extractor: a list of attributes if \code{name} is \code{NULL},
otherwise the value of the attribute matched, or NULL if no exact match is 
found and no or more than one partial match is found.
}
\description{
Query sf attributes from the active element of an sfnetwork object
}
\details{
sf attributes include \code{sf_column} (the name of the sf column)
and \code{agr} (the attribute-geometry-relationships).
}
