% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrs.R
\name{sf_attr}
\alias{sf_attr}
\title{Query sf attributes from the active element of a sfnetwork}
\usage{
sf_attr(x, name, active = NULL)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{name}{Name of the attribute to query. Either \code{'sf_column'} to
extract the name of the geometry list column, or \code{'agr'} to extract the
specification of attribute-geometry relationships.}

\item{active}{Which network element (i.e. nodes or edges) to activate before
extracting. If \code{NULL}, it will be set to the current active element of
the given network. Defaults to \code{NULL}.}
}
\value{
The value of the queried attribute.
}
\description{
Query sf attributes from the active element of a sfnetwork
}
\examples{
net = as_sfnetwork(roxel)
sf_attr(net, "agr", active = "edges")
sf_attr(net, "sf_column", active = "nodes")

}
