% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfnetwork.R
\name{sfnetwork}
\alias{sfnetwork}
\title{Create an sfnetwork object}
\usage{
sfnetwork(
  nodes,
  edges,
  directed = TRUE,
  edges_as_lines = NULL,
  force = FALSE,
  ...
)
}
\arguments{
\item{nodes}{An object containing information about the nodes in the network.
The nodes should contain geospatial coordinates, either by being an \code{sf}
object with \code{POINT} geometry features, or by being convertable to such an
object with \code{\link[sf]{st_as_sf}}.}

\item{edges}{An object containing information about the edges in the network.
This object may contain explicit geospatial information by being an \code{sf}
object with \code{LINESTRING} geometry features. However, this is optional. It
may also be a regular \code{data.frame} or \code{tbl_df} object. In any case,
the adjacent nodes of each edge must either be encoded in a \code{to} and
\code{from} column, or in the two first columns, as integers. These integers
refer to nodes indices, which in turn refer to the position of a node in the
nodes table.}

\item{directed}{Should the constructed network be directed? Defaults to
\code{TRUE}.}

\item{edges_as_lines}{Should the edges be spatially explicit, i.e. have
\code{LINESTRING} geometries stored in a geometry list column? If \code{NULL},
this will be automatically defined, by setting the argument to \code{TRUE}
when the given edges object contains a geometry list column, and \code{FALSE}
otherwise. Defaults to \code{NULL}.}

\item{force}{Should network validity checks be skipped? Defaults to
\code{FALSE}, meaning that network validity checks are executed when
constructing the network. These checks guarantee a valid spatial network
structure. For the nodes, this means that they all should have \code{POINT}
geometries. In the case of spatially explicit edges, it is also checked that
all edges have \code{LINESTRING} geometries, nodes and edges have the same
CRS and boundary points of edges match their corresponding node coordinates.
These checks are important, but also time consuming. If you are already sure
your input data meet the requirements, the checks are unneccesary and can be
turned off to improve speed.}

\item{...}{Arguments passed on to \code{\link[sf]{st_as_sf}}, when
converting the nodes to an \code{sf} object.}
}
\value{
An object of class \code{sfnetwork}.
}
\description{
\code{sfnetwork} is a tidy data structure for geospatial networks. It extends
the graph manipulation functionalities of the
\code{\link[tidygraph]{tidygraph-package}} package into the domain of
geospatial networks, where the nodes and optionally also the edges are
embedded in geographical space, and enables to apply the spatial analytical
function from the \code{\link[sf:sf]{sf-package}} directly to the network.
}
