% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfnetworks-package.R
\docType{package}
\name{sfnetworks-package}
\alias{sfnetworks}
\alias{sfnetworks-package}
\title{sfnetworks: Tidy Geospatial Networks}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a tidy approach to spatial network analysis, in the form of classes and functions that enable a seamless interaction between the network analysis package 'tidygraph' and the spatial analysis package 'sf'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://luukvdmeer.github.io/sfnetworks/}
  \item \url{https://github.com/luukvdmeer/sfnetworks}
  \item Report bugs at \url{https://github.com/luukvdmeer/sfnetworks/issues/}
}

}
\author{
\strong{Maintainer}: Lucas van der Meer \email{luukvandermeer@live.nl} (\href{https://orcid.org/0000-0001-6336-8628}{ORCID})

Authors:
\itemize{
  \item Lorena Abad \email{lore.abad6@gmail.com} (\href{https://orcid.org/0000-0003-0554-734X}{ORCID})
  \item Andrea Gilardi \email{andrea.gilardi@unimib.it} (\href{https://orcid.org/0000-0002-9424-7439}{ORCID})
  \item Robin Lovelace \email{r.lovelace@leeds.ac.uk} (\href{https://orcid.org/0000-0001-5679-6536}{ORCID})
}

}
\keyword{internal}
