% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth_pseudo_nodes}
\alias{smooth_pseudo_nodes}
\title{Smooth pseudo nodes}
\usage{
smooth_pseudo_nodes(
  x,
  protect = NULL,
  require_equal = NULL,
  attribute_summary = "ignore",
  store_original_ids = FALSE,
  store_original_data = FALSE
)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{protect}{An integer vector of edge indices specifying which nodes
should be protected from being removed. Defaults to \code{NULL}, meaning
that none of the nodes is protected.}

\item{require_equal}{A character vector of edge column names specifying
which attributes of the incident edges of a pseudo node should be equal in
order for the pseudo node to be removed? Defaults to \code{NULL}, meaning
that attribute equality is not considered for pseudo node removal.}

\item{attribute_summary}{How should the attributes of concatenated edges
be summarized? There are several options, see
\code{\link[igraph]{igraph-attribute-combination}} for details.}

\item{store_original_ids}{For each concatenated edge, should the indices of
the original edges be stored as an attribute of the new edge, in a column
named \code{.tidygraph_edge_index}? This is in line with the design
principles of \code{tidygraph}. Defaults to \code{FALSE}.}

\item{store_original_data}{For each concatenated edge, should the data of
the original edges be stored as an attribute of the new edge, in a column
named \code{.orig_data}? This is in line with the design principles of
\code{tidygraph}. Defaults to \code{FALSE}.}
}
\value{
The smoothed network as object of class \code{\link{sfnetwork}}.
}
\description{
Construct a smoothed version of the network by iteratively removing pseudo
nodes, while preserving the connectivity of the network. In the case of
directed networks, pseudo nodes are those nodes that have only one incoming
and one outgoing edge. In undirected networks, pseudo nodes are those nodes
that have two incident edges. Equality of attribute values among the two
edges can be defined as an additional requirement by setting the
\code{require_equal} parameter. Connectivity of the network is preserved by
concatenating the incident edges of each removed pseudo node.
}
