% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphers.R
\name{spatial_morphers}
\alias{spatial_morphers}
\alias{to_spatial_coordinates}
\alias{to_spatial_subdivision}
\alias{to_spatial_directed}
\alias{to_spatial_explicit_edges}
\alias{to_spatial_implicit_edges}
\alias{to_spatial_shortest_paths}
\alias{to_spatial_simple}
\alias{to_spatial_smooth}
\alias{to_spatial_subset}
\title{Spatial morphers for sfnetworks}
\usage{
to_spatial_coordinates(x)

to_spatial_subdivision(x)

to_spatial_directed(x)

to_spatial_explicit_edges(x, ...)

to_spatial_implicit_edges(x)

to_spatial_shortest_paths(x, ...)

to_spatial_simple(x, remove_parallels = TRUE, remove_loops = TRUE)

to_spatial_smooth(x, require_equal_attrs = FALSE)

to_spatial_subset(x, ..., subset_by = NULL)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{...}{Arguments to be passed on to other functions. See the description
of each morpher for details.}

\item{remove_parallels}{Should parallel edges be removed. Defaults to
\code{TRUE}.}

\item{remove_loops}{Should loops be remove. Defaults to \code{TRUE}.}

\item{require_equal_attrs}{Should pseudo nodes only be removed when all
attributes of their incident edges are equal? Defaults to \code{FALSE}.}

\item{subset_by}{Whether to create subgraphs based on nodes or edges.}
}
\value{
Either a \code{morphed_sfnetwork}, which is a list of one or more
\code{\link{sfnetwork}} objects, or a \code{morphed_tbl_graph}, which is a
list of one or more \code{\link[tidygraph]{tbl_graph}} object. See the
description of each morpher for details.
}
\description{
Spatial morphers form spatial add-ons to the set of
\code{\link[tidygraph]{morphers}} provided by \code{tidygraph}. These
functions are not meant to be called directly. They should either be passed
into \code{\link[tidygraph]{morph}} to create a temporary alternate
representation of the input network. Such an alternate representation is a
list of one or more network objects. Single elements of that list can be
extracted directly as a new network by passing the morpher to
\code{\link[tidygraph]{convert}} instead. Alternatively, if the morphed
state contains multiple elements, all of them can be extracted together
inside a \code{\link[tibble]{tibble}} by passing the morpher to
\code{\link[tidygraph]{crystallise}}.
}
\details{
It also possible to create your own morphers. See the documentation
of \code{\link[tidygraph]{morph}} for the requirements for custom morphers.
}
\section{Functions}{
\itemize{
\item \code{to_spatial_coordinates}: Store the spatial coordinates of the nodes in
separate coordinate columns, instead of a \code{\link[sf]{sfc}} geometry
list column. If edges are spatially explicit, the edge geometries are
dropped. Returns a \code{morphed_tbl_graph} containing a single element of
class \code{\link[tidygraph]{tbl_graph}}.

\item \code{to_spatial_subdivision}: Construct a subdivision of the network by
subdividing edges at each internal point that has an equal geometry to any
other internal or boundary point in the edges table. Internal points in this
sense are those points that are included in their linestring geometry
feature but are not endpoints of it, while boundary points are the endpoints
of the linestrings. The network is reconstructed after subdivision such that
edges are connected at the points of subdivision. Returns a
\code{morphed_sfnetwork} containing a single element of class
\code{\link{sfnetwork}}. This morpher requires edges to be spatially
explicit.

\item \code{to_spatial_directed}: Make a network directed in the direction given
by the linestring geometries of the edges. Differs from
\code{\link[tidygraph]{to_directed}}, which makes a network directed based
on the node indices given in the \code{from} and \code{to} columns. In
undirected networks these indices may not correspond with the endpoints of
the linestring geometries. Returns a \code{morphed_sfnetwork} containing a
single element of class \code{\link{sfnetwork}}. This morpher requires edges
to be spatially explicit. If not, use \code{\link[tidygraph]{to_directed}}.

\item \code{to_spatial_explicit_edges}: Create linestring geometries between from and
to nodes of spatially implicit edges. If the edges data can be directly
converted to an object of class \code{\link[sf]{sf}} using
\code{\link[sf]{st_as_sf}}, extra arguments can be provided as \code{...}
which will be forwarded to st_as_sf internally. Otherwise, straight lines
will be drawn between source and target nodes of edges. Returns a
\code{morphed_sfnetwork} containing a single element of class
\code{\link{sfnetwork}}.

\item \code{to_spatial_implicit_edges}: Remove linestring geometries of spatially
explicit edges. Returns a \code{morphed_sfnetwork} containing a single
element of class \code{\link{sfnetwork}}.

\item \code{to_spatial_shortest_paths}: Limit a network to those nodes and edges that
are part of the shortest path between two nodes. \code{...} is evaluated in
the same manner as \code{\link{st_network_paths}}. Returns a
\code{morphed_sfnetwork} that may contain multiple elements of class
\code{\link{sfnetwork}}, depending on the number of requested paths. When
unmorphing only the first instance of both the node and edge data will be
used, as the the same node and/or edge can be present in multiple paths.

\item \code{to_spatial_simple}: Remove loops and parallel edges. Returns a
\code{morphed_sfnetwork} containing a single element of class
\code{\link{sfnetwork}}. This morpher requires edges to be spatially
explicit. If not, use \code{\link[tidygraph]{to_simple}}.

\item \code{to_spatial_smooth}: Construct a smoothed version of the network by
iteratively removing pseudo nodes, while preserving the connectivity of the
network. In the case of directed networks, pseudo nodes are those nodes that
have only one incoming and one outgoing edge. In undirected networks, pseudo
nodes are those nodes that have two incident edges. Connectivity of the
network is preserved by concatenating the incident edges of each removed
pseudo node. Returns a \code{morphed_sfnetwork} containing a single element
of class \code{\link{sfnetwork}}.

\item \code{to_spatial_subset}: Subset the network by applying a spatial
filter, i.e. a filter on the geometry column based on a spatial predicate.
\code{...} is evaluated in the same manner as \code{\link{st_filter}}.
Returns a \code{morphed_sfnetwork} containing a single element of class
\code{\link{sfnetwork}}. For filters on an attribute column, use
\code{\link[tidygraph]{to_subgraph}}.
}}

\examples{
library(sf)
library(tidygraph)

net = as_sfnetwork(roxel, directed = FALSE) \%>\%
  st_transform(3035)

## to_spatial_coordinates
convert(net, to_spatial_coordinates)

## to_spatial_subdivision
# Number of edges in original network.
net \%>\%
  activate("edges") \%>\%
  st_as_sf() \%>\%
  nrow()
# Number of edges in subdivided network.
net \%>\%
  convert(to_spatial_subdivision) \%>\%
  activate("edges") \%>\%
  st_as_sf() \%>\%
  nrow()

## to_spatial_directed
net \%>\%
  convert(to_undirected) \%>\%
  convert(to_spatial_directed)

## to_spatial_explicit_edges
par(mar = c(1, 1, 1, 1), mfrow = c(1,2))
plot(net)
net \%>\%
 convert(to_spatial_implicit_edges) \%>\%
 convert(to_spatial_explicit_edges) \%>\%
 plot()

## to_spatial_implicit_edges
net \%>\%
 convert(to_spatial_implicit_edges)

## to_spatial_shortest_paths
# Plot shortest path.
par(mar = c(1, 1, 1, 1), mfrow = c(1,1))
plot(net)
net \%>\%
  convert(to_spatial_shortest_paths, 171, 190) \%>\%
  plot(col = "red", add = TRUE)
# Calculate lengths of multiple shortest paths.
net \%>\%
 activate("edges") \%>\%
 morph(to_spatial_shortest_paths, 1, c(171, 190)) \%>\%
 sapply(function(x) sum(st_length(x)))

## to_spatial_simple
# Number of edges in original network.
net \%>\%
  activate("edges") \%>\%
  st_as_sf() \%>\%
  nrow()
# Number of edges in simplified network.
net \%>\%
  activate("edges") \%>\%
  convert(to_spatial_simple) \%>\%
  st_as_sf() \%>\%
  nrow()

## to_spatial_smoothed
G = as_sfnetwork(roxel[c(4, 5, 8), ], directed = FALSE)
# Remove pseudo nodes.
smoothed_1 = convert(G, to_spatial_smooth)
# Only remove pseudo nodes when attributes of incident edges are equal.
smoothed_2 = convert(G, to_spatial_smooth, require_equal_attrs = TRUE)
# Compare results.
par(mar = c(1, 1, 1, 1), mfrow = c(1,3))
plot(G, cex = 3, main = "Original network")
plot(smoothed_1, cex = 3, main = "Smoothed network 1")
plot(smoothed_2, cex = 3, main = "Smoothed network 2")

## to_spatial_subset
p1 = st_point(c(7.53173, 51.95662))
p2 = st_point(c(7.53173, 51.95190))
p3 = st_point(c(7.53778, 51.95190))
p4 = st_point(c(7.53778, 51.95662))
rect = st_multipoint(c(p1, p2, p3, p4)) \%>\%
  st_cast('POLYGON') \%>\%
  st_sfc(crs = 4326) \%>\%
  st_transform(3035)
net \%>\%
  convert(to_spatial_subset, rect, .predicate = st_intersects)

}
