% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{spatial_node_types}
\alias{spatial_node_types}
\alias{node_is_pseudo}
\title{Query spatial node types}
\usage{
node_is_pseudo()
}
\value{
A logical vector of the same length as the number of nodes in the
network, indicating if each node is of the type in question.
}
\description{
These functions are a collection of node type queries that are commonly
used in spatial network analysis, and form a spatial extension to
\code{\link[tidygraph:node_types]{node type queries}} in tidygraph.
}
\details{
Just as with all query functions in tidygraph, these functions
are meant to be called inside tidygraph verbs such as
\code{\link[tidygraph]{mutate}} or \code{\link[tidygraph]{filter}}, where
the network that is currently being worked on is known and thus not needed
as an argument to the function. If you want to use an algorithm outside of
the tidygraph framework you can use \code{\link[tidygraph]{with_graph}} to
set the context temporarily while the algorithm is being evaluated.
}
\section{Functions}{
\itemize{
\item \code{node_is_pseudo()}: Pseudo nodes in directed networks are those
nodes with only one incoming and one outgoing edge. In undirected networks
pseudo nodes are those nodes with only two incident edges.

}}
