% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend.R
\name{st_blend}
\alias{st_blend}
\title{Blend geospatial points into a spatial network}
\usage{
st_blend(x, y, tolerance = Inf, sort = FALSE)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{y}{The spatial features to be blended, either as object of class
\code{\link[sf]{sf}} or \code{\link[sf]{sfc}}, with \code{POINT} geometries.}

\item{tolerance}{The tolerance distance to be used. Only features that are
at least as close to the network as the tolerance distance will be blended.
Should be a non-negative number preferably given as an object of class 
\code{\link[units]{units}}. Otherwise, it will be assumed that the unit is 
meters. If set to \code{Inf} all features will be blended. Defaults to 
\code{Inf}.}

\item{sort}{Should the nodes in the resulting network be sorted in the same 
order as those in x, followed by the new nodes blended in from y? Defaults 
to \code{FALSE}, meaning that node order might be changed. However, sorting
might influence performance.}
}
\value{
An object of class \code{\link{sfnetwork}}.
}
\description{
Blending a point into a network is the combined process of first snapping 
the given point to its nearest point on its nearest edge in the network, 
subsequently splitting that edge at the location of the snapped point, and
finally adding the snapped point as node to the network. If the location
of the snapped point is already a node in the network, the attributes of the
point (if any) will be joined to that node.
}
