% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{st_network_bbox}
\alias{st_network_bbox}
\title{Get the bounding box of a sfnetwork}
\usage{
st_network_bbox(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{...}{Arguments passed on to \code{\link[sf]{st_bbox}}.}
}
\value{
An object of class \code{bbox}.
}
\description{
A spatial network specific bounding box extractor, returning the combined
bounding box of the nodes and edges in the network.
}
\details{
See \code{\link[sf]{st_bbox}} for details.
}
