% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{st_network_join}
\alias{st_network_join}
\title{Join sfnetworks based on equality of node geometries}
\usage{
st_network_join(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{y}{An object of class \code{\link{sfnetwork}}.}

\item{...}{Arguments passed on to \code{\link[tidygraph]{graph_join}}.}
}
\value{
An object of class \code{\link{sfnetwork}}.
}
\description{
This spatial network specific join function makes a spatial full join on the 
geometries of the nodes data and updates the edges in y such that they
match the new indices of the nodes in the resulting network. Edge data is 
combined using a \code{bind_rows} semantic, meaning that data are matched 
by column name and filled with `NA` if it is missing in either of the
networks. Also, this means that two edges with equal geometries will both
be present in the resulting network.
}
