% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{st_round}
\alias{st_round}
\title{Rounding of geometry coordinates}
\usage{
st_round(x, digits = 0)
}
\arguments{
\item{x}{An object of class \code{\link[sf]{sf}} or \code{\link[sf]{sfc}}.}

\item{digits}{Integer indicating the number of decimal places to be used.}
}
\value{
An object of class \code{\link[sf]{sf}} or \code{\link[sf]{sfc}}
with rounded coordinates.
}
\description{
Rounding of geometry coordinates
}
\examples{
library(sf, quietly = TRUE)

p1 = st_sfc(st_point(c(1.123, 1.123)))
p2 = st_sfc(st_point(c(0.789, 0.789)))
p3 = st_sfc(st_point(c(1.123, 0.789)))

st_round(st_as_sf(c(p1, p2, p2, p3, p1)), digits = 1)

}
\seealso{
\code{\link{round}}
}
