% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygraph.R
\name{tidygraph}
\alias{tidygraph}
\alias{as_tbl_graph.sfnetwork}
\alias{as_tibble.sfnetwork}
\alias{convert.sfnetwork}
\alias{morph.sfnetwork}
\alias{mutate.sfnetwork}
\alias{select.sfnetwork}
\title{tidygraph methods for sfnetwork objects}
\usage{
\method{as_tbl_graph}{sfnetwork}(x, ...)

\method{as_tibble}{sfnetwork}(x, active = NULL, spatial = TRUE, ...)

\method{convert}{sfnetwork}(.data, ...)

\method{morph}{sfnetwork}(.data, .f, ...)

\method{mutate}{sfnetwork}(.data, ...)

\method{select}{sfnetwork}(.data, ...)
}
\arguments{
\item{x}{An object of class \code{\link{sfnetwork}}.}

\item{...}{Arguments passed on the corresponding \code{tidygraph} function.}

\item{active}{Which network element (i.e. nodes or edges) to activate before
extracting. If \code{NULL}, it will be set to the current active element of
the given network. Defaults to \code{NULL}.}

\item{spatial}{Should te extracted tibble be a 'spatial tibble', i.e. an
object of class \code{c('sf', 'tbl_df')}, if it contains a geometry list
column. Defaults to \code{TRUE}.}

\item{.data}{An object of class \code{\link{sfnetwork}}.}

\item{.f}{See \code{\link[tidygraph]{morph}}.}
}
\description{
\code{\link[tidygraph:tidygraph-package]{tidygraph}} methods for 
\code{\link{sfnetwork}} objects. Since \code{\link{sfnetwork}} objects
subclass \code{\link[tidygraph]{tbl_graph}} objects, most tidygraph
functions work automatically. However, some of them need a special method,
mostly as a result of the presence of the geometry list column in
\code{\link{sfnetwork}} objects. Use these methods without the .sfnetwork 
suffix and after loading the tidygraph package.
}
\details{
See the \code{\link[tidygraph:tidygraph-package]{tidygraph}} 
documentation.
}
\section{Functions}{
\itemize{
\item \code{as_tibble.sfnetwork}: The sfnetwork method for 
\code{\link[tidygraph]{as_tibble}} differs in the sense that whenever a
geometry list column is present, by default it will return a 
'spatial tibble'. With that we mean an object of class 
\code{c('sf', 'tbl_df')} instead of an object of class \code{'tbl_df'}.

\item \code{convert.sfnetwork}: The sfnetwork method for \code{\link[tidygraph]{convert}}
works the same, but updates the sf attributes of the resulting network.

\item \code{morph.sfnetwork}: The sfnetwork method for 
\code{\link[tidygraph]{morph}} will first try to input the 
\code{\link{sfnetwork}} object into the morph method for a 
\code{\link[tidygraph]{tbl_graph}}. If this fails, it will first convert
the \code{\link{sfnetwork}} object into a \code{\link[tidygraph]{tbl_graph}}
object before calling \code{\link[tidygraph]{morph}}. The returned value
will be a \code{morphed_sfnetwork} when all elements of the morphed graph
are of class \code{\link{sfnetwork}}, and a \code{morphed_tbl_graph}
otherwise.

\item \code{mutate.sfnetwork}: The sfnetwork method for \code{\link[tidygraph]{mutate}}
works the same, but updates the sf attributes of the resulting network.

\item \code{select.sfnetwork}: The sfnetwork method for \code{\link[tidygraph]{select}}
works the same, but updates the sf attributes of the resulting network.
}}

